% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHelper.R
\name{AddAbbreviationsSBxlsx}
\alias{AddAbbreviationsSBxlsx}
\alias{f_AddAbbreviationsSBxlsx}
\title{AddAbbreviationsSBxlsx}
\usage{
f_AddAbbreviationsSBxlsx(
  kaas = as_tibble(World$kaas),
  SubcompartmentsMap = c(marinesediment = "sd2", freshwatersediment = "sd1", lakesediment
    = "sd0", agriculturalsoil = "s2", naturalsoil = "s1", othersoil = "s3", air = "a",
    deepocean = "w3", sea = "w2", river = "w1", lake = "w0", cloudwater = "cw"),
  ScalesMap = c(Arctic = "A", Moderate = "M", Tropic = "T", Continental = "C", Regional =
    "R"),
  SpeciesMap = c(Dissolved = "D", Gas = "G", Large = "P", Small = "A", Solid = "S",
    Unbound = "U")
)
}
\arguments{
\item{kaas}{table with columns toSpecies, toSubCompart and toScale, idem from...}

\item{SubcompartmentsMap}{names character vector with the abreviations for each SubCompart}

\item{ScalesMap}{names character vector with the abreviations for each Scale}

\item{SpeciesMap}{names character vector with the abreviations for each Species}
}
\value{
Table with standard SimpleBox abbreviations for each compartment
}
\description{
Returns a table, default for World$kaas with two new columns to and from with the Abbrevaition of each SimpleBox compartment
}
