% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Adsorbtion.R
\name{Adsorption}
\alias{Adsorption}
\alias{k_Adsorbtion}
\title{Adsorption of chemical to across an interface}
\usage{
k_Adsorbtion(
  from.FRingas,
  from.FRinw,
  from.MTC_2sd,
  all.MTC_2w,
  from.MTC_2a,
  from.MTC_2s,
  FRorig,
  Kacompw,
  Kscompw,
  to.Matrix,
  VertDistance,
  landFRAC,
  from.SubCompart,
  to.SubCompart
)
}
\arguments{
\item{FRinw}{}

\item{MTC_2sd}{}

\item{Matrix}{# can use paraminherit to use description from another function! No need to copy paste same descriptions!}
}
\value{
The adsorption rate constant relevant for the receiving compartments soil, water or sediment [s-1]
}
\description{
The adsorption rate constant (k_Adsorption) is calculated based on the ... See 2015-0161 report.
}
