% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_AirFlow.R
\name{AirFlow}
\alias{AirFlow}
\title{AirFlow}
\usage{
AirFlow(Volume, Area, WINDspeed, SubCompartName)
}
\arguments{
\item{Volume}{volume in [m3]}

\item{Area}{area of the compartment in [m2]}

\item{WINDspeed}{within the compartment [m.s-1]}

\item{SubCompartName}{for which the calculation is executed}
}
\value{
AirFlow, from one scale to another
}
\description{
Computes the airflow within a compartment
}
