% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Area.R
\name{Area}
\alias{Area}
\title{Area}
\usage{
Area(
  AreaLand,
  AreaSea,
  landFRAC,
  all.landFRAC,
  SubCompartName,
  ScaleName,
  Regional_and_Continental_deepocean
)
}
\arguments{
\item{landFRAC}{fraction of land in the considered compartment [-]}

\item{SubCompartName}{the subcompartment which is considered}

\item{ScaleName}{the scale which is considered}

\item{AREAland}{the land area in the considered compartment [m2]}

\item{AREAsea}{the sea area in the considered compartment [m2]}
}
\value{
Area (but not for sediment) based on data for the SubCompartment / Scale
}
\description{
Area
}
