% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6CalcGraphModule.R
\name{CalcGraphModule}
\alias{CalcGraphModule}
\title{CalcGraphModule}
\description{
Mother of all Modules where all defining functions will be called
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{myName}}{the name of the defining function AND used to identify self}

\item{\code{exeFunction}}{r.o. property: the actual function that will be called}

\item{\code{myCore}}{r.o. property, see parameter TheCore at initialize}

\item{\code{needVars}}{parameter names of exefunction; used for the calculation graph}

\item{\code{FromAndTo}}{used for processes and flows; from which state(box) to what other state}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalcGraphModule-new}{\code{CalcGraphModule$new()}}
\item \href{#method-CalcGraphModule-fetch_current}{\code{CalcGraphModule$fetch_current()}}
\item \href{#method-CalcGraphModule-clone}{\code{CalcGraphModule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalcGraphModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalcGraphModule-new}{}}}
\subsection{Method \code{new()}}{
init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalcGraphModule$new(TheCore, exeFunction, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TheCore}}{the SBcore, parent object of self}

\item{\code{exeFunction}}{defining function, usually the name of, as character, 
for identification (a function does not know its own name)}

\item{\code{...}}{other parameters for various uses, end up in private$MoreParams}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalcGraphModule-fetch_current"></a>}}
\if{latex}{\out{\hypertarget{method-CalcGraphModule-fetch_current}{}}}
\subsection{Method \code{fetch_current()}}{
fetch specific values from core
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalcGraphModule$fetch_current(withoutValues)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{withoutValues}}{data.frame-ish with columns `varname` and needed D}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalcGraphModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalcGraphModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalcGraphModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
