% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPClassicNano.R
\name{ClassicNanoProcess}
\alias{ClassicNanoProcess}
\title{ClassicNanoProcess}
\description{
a process representing (all speed constants in) the classic nano version in excel
Creates functions that can be called from the World$ environment
}
\section{Super classes}{
\code{\link[sboo:CalcGraphModule]{sboo::CalcGraphModule}} -> \code{\link[sboo:ProcessModule]{sboo::ProcessModule}} -> \code{ClassicNanoProcess}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{r.o . Nano or molecular}

\item{\code{version}}{r.o. as given in the 1rst sheet of the file}

\item{\code{namedValues}}{relevant cells r.o. 
return tree with variable and its pedegree
#param cellName name (in excel) to look for
#param dependingOn description
#param as.DAG make a iGraph (default)? or as data.tree
#param maxDepth max. depth of recursion
return data.frame with variable in sb4n spreatsheet and their values
#param grepstr pattern
#param filename sb4n spreatsheet}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClassicNanoProcess-new}{\code{ClassicNanoProcess$new()}}
\item \href{#method-ClassicNanoProcess-Excelgrep}{\code{ClassicNanoProcess$Excelgrep()}}
\item \href{#method-ClassicNanoProcess-Exceltrace}{\code{ClassicNanoProcess$Exceltrace()}}
\item \href{#method-ClassicNanoProcess-Exceldependencies}{\code{ClassicNanoProcess$Exceldependencies()}}
\item \href{#method-ClassicNanoProcess-ExcelEmissions}{\code{ClassicNanoProcess$ExcelEmissions()}}
\item \href{#method-ClassicNanoProcess-ExcelSB.K}{\code{ClassicNanoProcess$ExcelSB.K()}}
\item \href{#method-ClassicNanoProcess-clone}{\code{ClassicNanoProcess$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sboo" data-topic="CalcGraphModule" data-id="fetch_current"><a href='../../sboo/html/CalcGraphModule.html#method-CalcGraphModule-fetch_current'><code>sboo::CalcGraphModule$fetch_current()</code></a></span></li>
<li><span class="pkg-link" data-pkg="sboo" data-topic="ProcessModule" data-id="execute"><a href='../../sboo/html/ProcessModule.html#method-ProcessModule-execute'><code>sboo::ProcessModule$execute()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoProcess-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoProcess-new}{}}}
\subsection{Method \code{new()}}{
init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoProcess$new(TheCore, filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TheCore}}{SBCore object}

\item{\code{filename}}{of the excel file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoProcess-Excelgrep"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoProcess-Excelgrep}{}}}
\subsection{Method \code{Excelgrep()}}{
find cells where the cellname conforms the grep, with their content
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoProcess$Excelgrep(grepstr, SenseCase = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grepstr}}{see grep}

\item{\code{SenseCase}}{is by default off, to comfort the novices}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoProcess-Exceltrace"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoProcess-Exceltrace}{}}}
\subsection{Method \code{Exceltrace()}}{
find cells depending on the initial cell with CellName, and their descendance
downto the maxDepth grandchildren; or vise-versa
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoProcess$Exceltrace(CellName, maxDepth = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CellName}}{starting point of the tree of descending cells}

\item{\code{maxDepth}}{recursion depth}

\item{\code{dependingOn}}{trace the origines for the cell(formula) or the dependencies (the default)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoProcess-Exceldependencies"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoProcess-Exceldependencies}{}}}
\subsection{Method \code{Exceldependencies()}}{
convenient function to find cells depending on the initial cell with CellName, 
and their descendance downto the maxDepth grandchildren
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoProcess$Exceldependencies(CellName, maxDepth = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{CellName}}{see Exceltrace}

\item{\code{maxDepth}}{see Exceltrace}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoProcess-ExcelEmissions"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoProcess-ExcelEmissions}{}}}
\subsection{Method \code{ExcelEmissions()}}{
read the emission data from the scenario sheet under the
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoProcess$ExcelEmissions(scenario_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scenario_name}}{column}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoProcess-ExcelSB.K"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoProcess-ExcelSB.K}{}}}
\subsection{Method \code{ExcelSB.K()}}{
read all the speed constants and their from~ and to box dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoProcess$ExcelSB.K()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoProcess-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoProcess-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoProcess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
