% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_D.R
\name{D}
\alias{D}
\title{D}
\usage{
D(FRorig, pKa, Kow, ChemClass)
}
\arguments{
\item{FRorig}{Fraction of species in its original state}

\item{pKa}{Dissociation constant of (conjugated) acid (default = 7)}

\item{Kow}{octanol water partitioning coefficient [-]}

\item{ChemClass}{Class of chemical; in this case Acid, Base or other}
}
\description{
octanol/water partitioning coefficient at neutral pH for colloids
}
