% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6EmissionModule.R
\name{EmissionModule}
\alias{EmissionModule}
\title{Emission module}
\description{
to set/get emissions, possibly make into functions
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmissionModule-new}{\code{EmissionModule$new()}}
\item \href{#method-EmissionModule-emissions}{\code{EmissionModule$emissions()}}
\item \href{#method-EmissionModule-emissionFunctions}{\code{EmissionModule$emissionFunctions()}}
\item \href{#method-EmissionModule-clone}{\code{EmissionModule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionModule-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionModule$new(emis, solvedAbbr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionModule-emissions"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionModule-emissions}{}}}
\subsection{Method \code{emissions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionModule$emissions(scenario = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionModule-emissionFunctions"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionModule-emissionFunctions}{}}}
\subsection{Method \code{emissionFunctions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionModule$emissionFunctions(emissions, states)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmissionModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmissionModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmissionModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
