% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Brown.R
\name{f_Brown}
\alias{f_Brown}
\title{Brownian motion or kinetic collision frequency coefficient (20181102)}
\usage{
f_Brown(Temp, viscosity, radius, radius_Otherparticle)
}
\arguments{
\item{Temp}{Temperature [K]}

\item{viscosity}{Dynamic viscosity of liquid (fraction of) compartment [kg.m-1.s-1]}

\item{radius}{Radius of nanoparticle [m]}

\item{radius_Otherparticle}{Radius of Other particle [m]}
}
\value{
fBrown [s-1]
}
\description{
Collision frequency of ENPs with other particulates due to Brownian motion (kinetic energy) in s-1 for heteroagglomeration. 
formula is based on Lyklema (2005)
}
