% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Diffusivity.R
\name{f_Diffusivity}
\alias{f_Diffusivity}
\title{Diffusivity}
\usage{
f_Diffusivity(Matrix, Temp, DynVisc, rad_species, Cunningham = NULL)
}
\arguments{
\item{Matrix}{as in c("air", "water", "soil")}

\item{Temp}{temperatures}

\item{DynVisc}{Dynamic viscosity (air, water)}

\item{rad_species}{see rad_species variable}

\item{Cunningham, }{see f_Cunningham}
}
\value{
Diffusivity
}
\description{
Diffusivity
}
