% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Inter.R
\name{f_Inter}
\alias{f_Inter}
\title{Interception or peri-kinetic collision frequency coefficient (20181102)}
\usage{
f_Inter(Shear = 0, from.radius, radius_Otherparticle)
}
\arguments{
\item{Shear}{Shear rate of the fluid matrix [s-1]}

\item{from.radius}{Radius of nanoparticle [m]}

\item{radius_Otherparticle}{Radius of Other particle [m]}
}
\value{
f_Inter [s-1]
}
\description{
Collission frequency of ENPs with other particulates due to fluid motion in s-1 for heteroagglomeration
}
