% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHelper.R
\name{f_MutateHelper}
\alias{f_MutateHelper}
\title{Using variable name and vector of values use mutateVars}
\usage{
f_MutateHelper(
  varName = "landFRAC",
  varData = c(0.6, 0.005, 0.27, 0.1, 0.025, 0.6, 0.005, 0.27, 0.1, 0.025)
)
}
\arguments{
\item{varName}{name of variable}

\item{varData}{vector of values to change to (in units for input!! use World$fetchDataUnits("landFRAC"), if unsure)}
}
\value{
mutated variable table
}
\description{
Applies the mutateVars sequence of data based on fetchData.
Make sure to set the correct vector of values, as not checks are done on value versus dimension!!!
}
