% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_PerimeterParticle.R
\name{f_PerimeterParticle}
\alias{f_PerimeterParticle}
\title{Calculate volume of perimeter of particle}
\usage{
f_PerimeterParticle(
  rad_particle,
  Shape,
  Longest_side = NULL,
  Intermediate_side = NULL,
  Shortest_side = NULL
)
}
\arguments{
\item{rad_particle}{Radius of particle [m]}

\item{Shape}{Particle Shape defined by user, used for nano and plastic [-]}

\item{Longest_side}{Longest side length as defined by user [m]}

\item{Intermediate_side}{Intermediate side length as defined by user [m]}

\item{Shortest_side}{Shortest side length as defined by user [m]}
}
\value{
fPerimeterParticle [m]
}
\description{
Calculate the perimeter of the 2d object of a 3d particle for ENPs
}
