% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_RadRain.R
\name{f_RadRain}
\alias{f_RadRain}
\title{Estimate radius of rain drops}
\usage{
f_RadRain(RAINrate, FRACtwet)
}
\arguments{
\item{RAINrate}{Yearly Rain rate [m/s]}

\item{FRACtwet}{Time fraction of wet episodes (~0-1) [-]}
}
\value{
Rainrate in wet episodes is added in SB4.0, for nanoparticles rain rate is used as in model definition prototype (average in whole year)
}
\description{
Estimate radius of rain drops
}
