% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_SchmidtNumber.R
\name{f_SchmidtNumber}
\alias{f_SchmidtNumber}
\title{SchmidtNumber}
\usage{
f_SchmidtNumber(
  visc,
  rhoMatrix,
  Diffusivity = NULL,
  Temp = NULL,
  rad_particle = NULL,
  Cunningham = NULL
)
}
\arguments{
\item{visc}{}

\item{rhoMatrix}{}

\item{Temp}{}

\item{rad_particle}{}
}
\value{
SchmidtNumber
}
\description{
a description follows .. Joris?
}
