% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_SetVelSolver.R
\name{f_SetVelSolver}
\alias{f_SetVelSolver}
\title{Settling Velocity Solver based on RSS}
\usage{
f_SetVelSolver(
  d_eq,
  Psi,
  DynViscFluidStandard,
  rhoParticle,
  rhoFluid,
  DragMethod,
  CSF,
  Matrix,
  rad_species,
  kS,
  kN
)
}
\arguments{
\item{d_eq}{Equivalent spherical diameter of the particle [-]}

\item{Psi}{Shape factor, circularity/sphericity [-]}

\item{DynViscFluidStandard}{Dynamic viscosity of liquid  (water or air) compartment [kg.m-1.s-1]}

\item{rhoParticle}{Density of nanoparticle [kg.m-3]}

\item{rhoFluid}{Density of water or air [kg m-3]}

\item{DragMethod}{Method of calculating the Drag Coefficient}

\item{CSF}{Corey Shape Factor [-]}

\item{rad_species}{radius of the species [m]}

\item{CD}{Drag Coefficient of a particle [-]}

\item{Re}{Reynolds number, as returned by the solver [-]}

\item{GN}{gravitational force constant [m2 s-1]}

\item{Cunningham}{Cunningham coefficient, see f_Cunningham [-]}
}
\value{
settling velocity [m/s]
}
\description{
Calculates the settling velocity by minimizing  the residual sum of squares (RSS)
}
