% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_SetVelWater.R
\name{f_SetVelWater}
\alias{f_SetVelWater}
\title{Function for calculating the Settling Velocity of a particle through water}
\usage{
f_SetVelWater(
  Shortest_side,
  rho_species,
  rhoMatrix,
  DynViscWaterStandard,
  DynViscAirStandard,
  Matrix,
  SubCompartName,
  Shape,
  Longest_side,
  Intermediate_side,
  DragMethod
)
}
\arguments{
\item{DynViscWaterStandard}{Dynamic viscosity of water []}

\item{rhoParticle}{Density of particle [kg/m3]}

\item{rhoWater}{Density of fluid matrix in which particle is present [kg/m3]}

\item{radius}{Radius of the particle [m]}

\item{GN}{gravitational force constabt [m2/s]}
}
\value{
f_SetVelWater
}
\description{
Settling Velocity or terminal Velocity of particles in a fluid medium based on Stokes' theorem (1901)
NB SettlingVelocity is a function to be used for all types of partical species, not just nano-species. Therefor it's not a variable defining function.
}
