% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_SurfaceAreaParticle.R
\name{f_SurfaceArea}
\alias{f_SurfaceArea}
\title{Calculate volume of sphere assuming (20181102)}
\usage{
f_SurfaceArea(
  Shape,
  Longest_side = NULL,
  Intermediate_side = NULL,
  Shortest_side = NULL,
  rad_particle
)
}
\arguments{
\item{Shape}{Particle Shape defined by user, used for nano and plastic [-]}

\item{Longest_side}{Longest side length as defined by user [m]}

\item{Intermediate_side}{Intermediate side length as defined by user [m]}

\item{Shortest_side}{Shortest side length as defined by user}

\item{rad_particle}{Radius of particle [m]}
}
\value{
f_SurfaceArea [m2]
}
\description{
Calculate the volume of spherical particle using the radius in m3, Shapes based on Gov4Nano
}
