% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_TAU.R
\name{f_TAU}
\alias{f_TAU}
\title{Calculate residence time}
\usage{
f_TAU(Area, WINDspeed)
}
\arguments{
\item{Area}{area of the compartment [m2]}

\item{WINDspeed}{windspeed within the compartment [m/s]}
}
\value{
TAU [s]
}
\description{
general approximation of mean residence time
}
