% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Cunningham.R
\name{fCunningham}
\alias{fCunningham}
\alias{f_Cunningham}
\title{Cunningham}
\usage{
f_Cunningham(rad_species)
}
\arguments{
\item{rad_species}{radius of the heteroagglomerates, calculated in v_rad_species [m]}
}
\value{
Cunningham [-]
}
\description{
Collision frequency of ENPs with other particulates due to Gravitational or Inertial impaction for heteroagglomeration.
Experimental values for the empirical constants Alpha (1.142), Beta (0.558), and Gamma (0.999) from Allan and Raab 1985 (Schneider and Voigt H2)
The default for "mean free path in air" is 66*10^-92, based on Seinfeld, and Pandis (2006)
}
