% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_FRinaerw.R
\name{fFRinearw}
\alias{fFRinearw}
\title{Fraction of chemical in aerosol water}
\usage{
fFRinearw(Kaers, Kaerw, FRACaer)
}
\arguments{
\item{FRACaer}{Fraction aerosols in air [-]}
}
\value{
The fraction of a chemical in the aerosol water phase. Total: FRingas + FRinaerw + FRinaers = 1.
}
\description{
Calculates the fraction of a chemical in aerosol water relative to the total amount of the chemical in air.
}
\seealso{
[Fringas(), FRinw(), FRins()]
}
