% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Grav.R
\name{fGrav}
\alias{fGrav}
\alias{f_Grav}
\title{Gravitational impaction collision frequency coefficient (20181102)}
\usage{
f_Grav(
  radiusParticle,
  rhoParticle,
  radius_Otherparticle,
  rho_Otherparticle,
  rhoFluid,
  DynViscWaterStandard
)
}
\arguments{
\item{radius_Otherparticle}{Radius of Other particle [m]}

\item{rho_Otherparticle}{Density (specific weight) of natural particle [kg/m3]}

\item{DynViscWaterStandard}{Dynamic viscosity of liquid  (fraction of) compartment [kg.m-1.s-1]}

\item{radius}{Radius of nanoparticle [m]}

\item{rho}{Density of nanoparticle [kg.m-3]}
}
\value{
fGrav [s-1]
}
\description{
Collission frequency of ENPs with other particulates due to Gravitational or Intertial impaction in s-1 for heteroagglomeration
}
