% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6FlowModule.R
\name{FlowModule}
\alias{FlowModule}
\title{R6 Flow module}
\description{
the module computing all flows between compartments.
}
\section{Super class}{
\code{\link[sboo:CalcGraphModule]{sboo::CalcGraphModule}} -> \code{FlowModule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlowModule-new}{\code{FlowModule$new()}}
\item \href{#method-FlowModule-execute}{\code{FlowModule$execute()}}
\item \href{#method-FlowModule-clone}{\code{FlowModule$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sboo" data-topic="CalcGraphModule" data-id="fetch_current"><a href='../../sboo/html/CalcGraphModule.html#method-CalcGraphModule-fetch_current'><code>sboo::CalcGraphModule$fetch_current()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-FlowModule-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowModule$new(TheCore, exeFunction, WithProcess)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowModule-execute"></a>}}
\if{latex}{\out{\hypertarget{method-FlowModule-execute}{}}}
\subsection{Method \code{execute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowModule$execute(debugAt = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlowModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
