% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_NumConc.R
\name{fNumConc}
\alias{fNumConc}
\alias{f_NumConc}
\title{Calculate number concentration from mass concentration}
\usage{
f_NumConc(rad_particle, rho_particle, MasConc)
}
\arguments{
\item{rad_particle}{Radius of particle [m]}

\item{MasConc}{Mass based concentration of particle per volume [kg.m-3]}

\item{rho_partocle}{Density of particle [kg.m-3]}
}
\value{
fNumConc [#.m-3]
}
\description{
Calculate number concentration from mass concentration [#.m-3]
}
