% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRACa.R
\name{FRACa}
\alias{FRACa}
\title{FRACa}
\usage{
FRACa(subFRACa, subFRACw, subFRACs, Matrix)
}
\arguments{
\item{subFRACa}{subfraction of air in a non-air compartment [-]}

\item{subFRACw}{subfraction of water in a non-water compartment [-]}

\item{subFRACs}{subfraction of solids in a non-soil, non-sediment compartment [-]}

\item{Matrix}{type of compartment}
}
\value{
FRACa
}
\description{
Calculates the total air fraction in the considered system.
either subFRACa or, when the main matrix, remainder after substracting subFRACs + subFRACw
}
