% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRACs.R
\name{FRACs}
\alias{FRACs}
\title{FRACs}
\usage{
FRACs(subFRACa, subFRACw, subFRACs, Matrix)
}
\arguments{
\item{subFRACa}{fraction of air in a non-air subcompartment [-]}

\item{subFRACw}{fraction of water in a non-water subcompartment [-]}

\item{subFRACs}{fraction of solids in a non-soil, non-sediment compartment [-]}

\item{Matrix}{type of compartment}
}
\value{
FRACs
}
\description{
fraction of solid in any matrix 
either subFRACs or, when the main matrix, remainder after substracting subFRACa + subFRACw
}
