% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRACa.R
\name{FRaction of air in any matrix}
\alias{FRaction of air in any matrix}
\alias{FRACa}
\title{FRACa}
\usage{
FRACa(subFRACa, subFRACw, subFRACs, Matrix)
}
\description{
either subFRACa or, when the main matrix, remainder after substracting subFRACs + subFRACw
}
