% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRACs.R
\name{FRaction of solid in any matrix}
\alias{FRaction of solid in any matrix}
\alias{FRACs}
\title{FRACs}
\usage{
FRACs(subFRACa, subFRACw, subFRACs, Matrix)
}
\description{
either subFRACs or, when the main matrix, remainder after substracting subFRACa + subFRACw
}
