% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRACw.R
\name{FRaction of water in any matrix}
\alias{FRaction of water in any matrix}
\alias{FRACw}
\title{FRACw}
\usage{
FRACw(subFRACa, subFRACw, subFRACs, Matrix)
}
\description{
either subFRACw or, when the main matrix, remainder after substracting subFRACs + subFRACa
}
