% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRinaerw.R
\name{FRinaerw}
\alias{FRinaerw}
\title{Fraction of chemical in aerosol water}
\usage{
FRinaerw(Kaers, Kaerw, FRACw, FRACs)
}
\arguments{
\item{Kaers}{Dimensionless aerosol solid / air Partitioning Coefficient [-]}

\item{Kaerw}{Dimensionless aerosol water / air Partitioning Coefficient [-]}

\item{FRACw}{fraction of water in the matrix [-]}

\item{FRACs}{fraction of solids in the matrix [-]}
}
\value{
The fraction of a chemical in the aerosol water phase. Total: FRingas + FRinaerw + FRinaers = 1.
}
\description{
Calculates the fraction of a chemical in aerosol water relative to the total amount of the chemical in air.
}
\seealso{
[Fringas(), FRinw(), FRins()]
}
