% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRinw.R
\name{Frinw}
\alias{Frinw}
\alias{FRinw}
\title{FRaction of chemical truely dissolved in water phase fresh water (relevant to Molecular species)}
\usage{
FRinw(
  FRorig_spw,
  FRACw,
  FRACa,
  FRACs,
  Kp,
  RHOsolid,
  KpCOL,
  Kacompw,
  SUSP,
  COL,
  Matrix
)
}
\arguments{
\item{D}{, to calculate Kp.col}
}
\description{
not the part suspended or colloidal
}
