% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRorig.R
\name{FRorig}
\alias{FRorig}
\title{Fraction of orginigal species - air, soil, sediment and water}
\usage{
FRorig(ChemClass, Matrix, pH, pKa)
}
\arguments{
\item{ChemClass}{Class of chemical, in this case Acid, Base or other}

\item{Matrix}{type of compartment considered}

\item{pH}{pH of soil, sediment, water or aerosol water}

\item{pKa}{Dissociation constant of (conjugated) acid (default = 7)}
}
\value{
FRorig
}
\description{
Fraction original species in air, soil, sediment or water
}
