% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRorig_spw.R
\name{FRorig_spw}
\alias{FRorig_spw}
\title{Fraction of original species - soil pore water}
\usage{
FRorig_spw(ChemClass, Matrix, pH, pKa)
}
\arguments{
\item{ChemClass}{Class of chemical, in this case Acid, Base or other}

\item{Matrix}{type of compartment considered}

\item{pH}{pH of soil}

\item{pKa}{Dissociation constant of (conjugated) acid (default = 7)}
}
\description{
Fraction original species in soil pore water
}
