% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_TAU.R
\name{fTAU}
\alias{fTAU}
\alias{f_TAU}
\title{Calculate residence time}
\usage{
f_TAU(Area, WINDspeed)
}
\arguments{
\item{Area}{in m2}

\item{WINDspeed}{in m s-1}
}
\value{
TAU [s]
}
\description{
general approximation of mean residence time
}
