% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Vol.R
\name{fVol}
\alias{fVol}
\title{Calculate volume of sphere or other shape}
\usage{
fVol(
  rad_particle = NULL,
  Shape = NULL,
  Longest_side = NULL,
  Intermediate_side = NULL,
  Shortest_side = NULL
)
}
\arguments{
\item{rad_particle}{Radius of particle [m]}

\item{Shape}{Particle shape defined by user, used for nano and plastic [-]}

\item{Longest_side}{Longest side length as defined by user [m]}

\item{Intermediate_side}{Intermediate side length as defined by user [m]}

\item{Shortest_side}{Shortest side length as defined by user}
}
\value{
fVol [m3]
}
\description{
Calculate the volume of spherical particle using the radius in m3, shapes based on Gov4Nano
}
