% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_IntrcptFrac.R
\name{IntrcptFrac}
\alias{IntrcptFrac}
\title{Fraction of intercepted particles in soil from Runoff}
\usage{
IntrcptFrac(VegInterceptFrac, SizeRunoff, rad_species, Matrix)
}
\arguments{
\item{Matrix}{type of compartment}

\item{subFRACa}{fraction of air in a non-air subcompartment [-]}

\item{subFRACw}{fraction of water in a non-water subcompartment [-]}

\item{subFRACs}{fraction of solids in a non-soil, non-sediment compartment [-]}
}
\value{
FRACs
}
\description{
Data from the interception experiments in Runoff from Han et al. 2022 are used to
}
