% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGeneral.R
\name{is.df.with}
\alias{is.df.with}
\title{Title}
\usage{
is.df.with(df, callingName, mustHavecols, reallystop = T)
}
\arguments{
\item{df}{input object df, or df-alike}

\item{callingName}{name of the calling function, or? user informative string}

\item{mustHavecols}{column names that the df should have}
}
\value{
side effect (stop) only
}
\description{
Title
}
