% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_HeteroAgglomeration.a.R
\name{k.HeteroAgglomeration.a}
\alias{k.HeteroAgglomeration.a}
\alias{k_HeteroAgglomeration.a}
\title{Heteroagglomeration with aerosols}
\usage{
k_HeteroAgglomeration.a(
  rad_species,
  RadOther,
  rho_species,
  RhoOther,
  Temp,
  Diffusivity,
  NumConcNuc,
  NumConcAcc,
  NumConcCP
)
}
\arguments{
\item{Temp}{Temperature of compartment [K]}

\item{to.alpha}{Attachment Efficiency of ENPs with other particulates [-]}

\item{NumConc_Otherparticle}{Mass concentration of other particulates [kg.m-3]}

\item{RadS}{Radius of nanoparticle [m]}

\item{RhoS}{Density of nanoparticle [kg.m-3]}

\item{radius_Otherparticle}{Radius of natural particle [m]}

\item{rho_Otherparticle}{Density (specific weight) of natural particle [kg/m3]}

\item{rhoMatrix}{Density of fluid matrix (water) [kg/m3]}

\item{viscosity}{Dyanic viscosity of fluid matrix []}
}
\value{
k.HeteroAgglomeration, the rate constant for 1rst order process: heteroagglomeration [s-1]
}
\description{
Calculation of the first order rate constant (s-1) for heteroagglomeration of ENPs with large (>0.45 um) natural particles
}
