% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_HeteroAgglomeration.sd.R
\name{k.HeteroAgglomeration.sd}
\alias{k.HeteroAgglomeration.sd}
\alias{k_HeteroAgglomeration.sd}
\title{Formation of heteroagglomerate (P) of ENPs (S) and solid grains in soil or sediment}
\usage{
k_HeteroAgglomeration.sd(
  Udarcy,
  rad_species,
  rho_species,
  to.RadOther,
  to.FRACs,
  to.alpha,
  Temp,
  hamakerSP.w
)
}
\arguments{
\item{Udarcy}{Darcu velocity in soil or sediment [m.s-1]}

\item{rad_species}{Radius (of nanoparticle) [m]}

\item{rho_species}{Density (of nanoparticle) [kg.m-3]}

\item{to.RadOther}{Radius of natural particle associated with the "to" species [m]}

\item{Temp}{Temperature of Scale [K]}

\item{hamakerSP.w}{Hamaker constant heteroagglomerate (ENP, water, SiO2) [J]}

\item{alpha}{Attachment Efficiency of ENPs with other particulates [-]}

\item{FRACs}{Fraction solid particles in soil or sediment [-]}
}
\value{
k.HeteroAgglomeration, the rate constant for 1rst order process: heteroagglomeration [s-1]
}
\description{
Calculation of the first order rate constant (s-1) for heteroagglomeration of ENPs with large (>0.45 um) natural particles
}
