% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_AdvectionRiverSeaScales.R
\name{k_AdvectionRiverSeaScales}
\alias{k_AdvectionRiverSeaScales}
\title{k_AdvectionRiverSeaScales}
\usage{
k_AdvectionRiverSeaScales(
  from.SubCompartName,
  from.ScaleName,
  to.ScaleName,
  from.Volume,
  to.Volume,
  ContRiver2Reg,
  ContSea2Reg,
  RegSea2Cont,
  OceanCurrent,
  from.TAUsea,
  to.TAUsea,
  ContinentalInModerate
)
}
\arguments{
\item{from.SubCompartName}{dimension dependency in function}

\item{from.ScaleName}{dimension dependency in function}

\item{to.ScaleName}{dimension dependency in function}

\item{from.Volume}{Volume [m3] of the from box}

\item{to.Volume}{Volume [m3] of the from box}

\item{ContRiver2Reg}{riverflow from continental to regional}

\item{ContSea2Reg}{seaflow from continental to regional}

\item{RegSea2Cont}{seaflow from regional to continental}

\item{OceanCurrent}{global constant}

\item{from.TAUsea}{refresh rate}

\item{to.TAUsea}{dito for the "to" box}

\item{ContinentalInModerate}{logical; continental can be in Moderate or in Tropic}
}
\value{
Advection seas and deepoceans, excluding vertical mixing sea <-> deepocean
}
\description{
k_AdvectionRiverSeaScales
}
