% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_AdvectionSeaOcean.R
\name{k_AdvectionSeaOcean}
\alias{k_AdvectionSeaOcean}
\title{k_AdvectionSeaOcean}
\usage{
k_AdvectionSeaOcean(
  to.Volume,
  from.Volume,
  to.TAUsea,
  from.TAUsea,
  OceanCurrent,
  SubCompartName,
  ScaleName
)
}
\arguments{
\item{to.Volume}{Volume of the to box}

\item{from.Volume}{Volume of the from box}

\item{to.TAUsea}{TAU - Residence time in sea; scale variable}

\item{from.TAUsea}{TAU - Residence time in sea; for the "to" box}

\item{OceanCurrent}{[m3.s-1]}

\item{SubCompartName}{name of the subcompartment of the box at hand}

\item{ScaleName}{name of the scale of the box at hand}
}
\value{
Advection sea - deepocean
}
\description{
k_AdvectionSeaOcean
}
