% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_AdvectionWaters.R
\name{k_AdvectionWaters}
\alias{k_AdvectionWaters}
\title{k_AdvectionWaters}
\usage{
k_AdvectionWaters(
  WATERflow.w0R.w2R,
  WATERflow.w1R.w0R,
  Volume,
  LakeOutflow,
  RiverDischarge
)
}
\arguments{
\item{WATERflow.w0R.w2R}{}

\item{WATERflow.w1R.w0R}{}

\item{Volume}{of box, SB variable}

\item{LakeOutflow}{flow}

\item{RiverDischarge}{SB variable}
}
\value{
Advection lake - river - to sea; only local scales
}
\description{
k_AdvectionWaters
}
