% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_CWscavenging.R
\name{k_CWscavenging}
\alias{k_CWscavenging}
\title{Cloud water scavenging coefficients as function of particle size and density}
\usage{
k_CWscavenging(
  RAINrate,
  FRACtwet,
  tdry,
  twet,
  COLLECTeff,
  rad_species,
  rho_species,
  to.rhoMatrix,
  from.rhoMatrix,
  DynViscAirStandard,
  DynViscWaterStandard,
  SettlingVelocity,
  Temp,
  Matrix,
  SpeciesName,
  VertDistance,
  SubCompartName
)
}
\arguments{
\item{RAINrate}{Yearly rain rate [m/s]}

\item{FRACtwet}{Time fraction of wet episodes [-]}

\item{tdry}{duration of dry episodes [s]}

\item{twet}{duration of wet episodes [s]}

\item{COLLECTeff}{Aerosol collection efficiency for raindrops (Wang et al., 2010) [-]}

\item{rad_species}{particle radius [m]}

\item{rho_species}{particle density [kg m-3]}

\item{to.rhoMatrix}{density of cloud water [kg m-3]}

\item{DynViscAirStandard}{dynamic viscosity of air [kg m-3 s-1]}

\item{DynViscWaterStandard}{dynamic viscosity of (cloud) water [kg m-3 s-1]}

\item{SettlingVelocity}{settling velocity for particle species in air [m/s]}

\item{Temp}{temperature [K]}

\item{Matrix}{compartment type}

\item{SpeciesName}{name of the considered species}

\item{VertDistance}{depth of compartment [m]}

\item{SubCompartName}{considered subcompartment}

\item{rad_Raindrop}{radius of the raindrop [m]}

\item{rhoMatrix}{density of air [kg m-3]}

\item{CritStokesNumb.cw}{Critical stokes number, based on Wang et al. (2010)}

\item{ReyNumber.cw}{Reynolds number}

\item{SchmidtNumber.Particle.a}{Denotes the ratio of momentum diffusivity and mass diffusivity, based on Wang et al. (2010)}
}
\value{
k_CWscavenging
}
\description{
Process describing the transport from dry air to rain, based on Wang et al. (2010) https://doi.org/10.5194/acp-10-5685-2010
}
