% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Deagglomeration.R
\name{k_Deagglomeration}
\alias{k_Deagglomeration}
\title{Fragmentation and degradation of plastics}
\usage{
k_Deagglomeration(
  kdeag,
  SubCompartName,
  ScaleName,
  Regional_and_Continental_deepocean
)
}
\arguments{
\item{kdeag}{Falling appart or fragmenting of heteroaglomerates [s-1]}

\item{Regional_and_Continental_deepocean}{If this variable is TRUE, Regional and Continental deepocean compartments are removed}
}
\value{
k_Deagglomeration, the combined degradation and fragmentation rate of microplastics [s-1]
}
\description{
Deagglomeration is the opposite of heteroagglomeration. For instance Polymer particles with an inorganic part can fragment. It is used in Plastics World. For instance for fragmentation of Tyre and Road wear particles into Tyre wear particles alone.
}
