% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Deposition.R
\name{k_Deposition}
\alias{k_Deposition}
\title{k_Deposition}
\usage{
k_Deposition(
  FRingas,
  WINDspeed,
  VertDistance,
  twet,
  tdry,
  COLLECTeff,
  AEROSOLdeprate,
  Kacompw,
  FRorig,
  SpeciesName,
  OtherkAir,
  landFRAC,
  Kaers,
  Kaerw,
  FRACaer
)
}
\arguments{
\item{VertDistance}{Mixing depth air compartment #[m]}

\item{FRorig}{Fraction original species in water or porewater}

\item{SpeciesName}{Name of the species the function is applied to (now only implemented for molecular, particulate drydep to be added) [NA]}

\item{OtherkAir}{correction term for processes other than deposition affecting the air compartment}

\item{FRaers}{}

\item{AREAFRAC}{Fraction of area compartment relative to the whole surface area of this scale (systemarea) [-]}

\item{RAINrate}{Average precipitation #[m/s]}
}
\value{
Transfer rate constant gaseous species air to soil or water #[s-1]
}
\description{
k_Deposition
}
