% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_DryDeposition.R
\name{k_DryDeposition}
\alias{k_DryDeposition}
\title{Dry deposition from air to surface soil or water}
\usage{
k_DryDeposition(
  to.Area,
  from.Volume,
  AEROresist,
  Diffusivity,
  DynVisc,
  rhoMatrix,
  ColRad,
  rad_species,
  rho_species,
  Temp,
  to.alpha.surf,
  Cunningham,
  SetVel
)
}
\arguments{
\item{to.Area}{Surface area of receiving land or water compartment [m2]}

\item{from.Volume}{Volume of air compartment [m3]}

\item{AEROresist}{aerodynamic resistance (Constant set to 74) [s.m-1]}

\item{Diffusivity}{TODO}

\item{DynVisc}{TODO}

\item{rhoMatrix}{Density of air [kg.m-3]
#param from.viscosity Dynamic viscosity of air compartment [kg.m-1.s-1]}

\item{ColRad}{Land surface particle collector radius [m] appendix A of LOTEUR.
#param FricVel Friction velocity [m s-1] (19 according to Van Jaarsveld, table 2.2)}

\item{rad_species}{rad_particle}

\item{rho_species}{rho_particle}

\item{Temp}{Temperature [K]}

\item{to.alpha.surf}{depends on the vegetation type, see table A.1 in ref guide LOTEUR v2.0 2016}

\item{Cunningham}{TODO}

\item{SetVel}{TODO}
}
\value{
k_drydeposition, the rate constant for 1rst order process: dry deposition from air to soil or water [s-1]
}
\description{
Calculation of the first order rate constant for deposition from air to the soil or water surface [s-1]
}
