% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Erosion.R
\name{k_Erosion}
\alias{k_Erosion}
\title{Erosion of particles}
\usage{
k_Erosion(
  relevant_depth_s,
  penetration_depth_s,
  EROSIONsoil,
  VertDistance,
  ScaleName,
  to.SubCompartName,
  Matrix
)
}
\arguments{
\item{EROSIONsoil}{Soil erosion #[mm/yr]}

\item{VertDistance}{Mixing depth soil #[m]}

\item{ScaleName}{To adjust for the absence of rivers in global scales}

\item{to.SubCompartName}{name of the subcompartment of the destination box of this process}

\item{CORRrunoff}{correction for runoff}
}
\value{
k_Erosion Erosion of attached enp species (P) from soil to water #[s-1]
}
\description{
Erosion of particles
}
