% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Escape.R
\name{k_Escape}
\alias{k_Escape}
\title{escape}
\usage{
k_Escape(t_half_Escape, to.SubCompartName, from.SubCompartName)
}
\arguments{
\item{t_half_Escape}{Half life time in air [s]}

\item{SubCompartName}{considered subcompartment}
}
\value{
k_Escape
}
\description{
calculate k for escape from air compartment to stratosphere based on t_half_Escape
}
