% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_DEGradation.R
\name{k_KDEGradation}
\alias{k_KDEGradation}
\alias{k_DEGradation}
\title{k_KDEGradation}
\usage{
k_DEGradation(
  FRingas,
  kdeg.air,
  C.OHrad.n,
  C.OHrad,
  k0.OHrad,
  Ea.OHrad,
  Temp,
  T25,
  kdeg.soil,
  kdeg.sediment,
  Q.10,
  kdeg.water,
  FRinw,
  BACTtest,
  BACTcomp,
  Matrix,
  SpeciesName
)
}
\arguments{
\item{FRingas}{}

\item{Temp}{[K]}

\item{T25}{= 298 K}
}
\value{
KDEGradationAir
}
\description{
calculate k for degradation in air;
}
