% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Resuspension.R
\name{k_Resuspenion}
\alias{k_Resuspenion}
\alias{k_Resuspension}
\title{Resuspension rate constant of substances in sediment}
\usage{
k_Resuspension(
  VertDistance,
  DynViscWaterStandard,
  from.rhoMatrix,
  to.NETsedrate,
  to.RadCP,
  to.RhoCP,
  from.RhoCP,
  FRACs,
  to.SUSP,
  SpeciesName
)
}
\arguments{
\item{VertDistance}{mixed depth sediment compartment #[m]}

\item{FRACs}{Volume fraction solids in sediment #[-]}

\item{to.SettlingVelocity}{SettlingVelocity, see function}

\item{RhoCP}{Mineral density of sediment and soil #[kg/m3]}

\item{to.mConcSusp}{Concentration of suspended matter in water #[kg/m3]}
}
\value{
k_Resuspension Resuspension flow from sediment #[s-1]
}
\description{
Calculation of the resuspension rate constant from sediment to water
For molecular substanes. More documentation to be added.
}
