% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_WetDeposition.R
\name{k_WetDeposition}
\alias{k_WetDeposition}
\title{Rate constant for wet deposition of particulate species due to rain}
\usage{
k_WetDeposition(to.Area, from.Volume, RAINrate)
}
\arguments{
\item{RAINrate}{average rain rate [m s-1]}

\item{to.AREA}{Surface area of receiving land or water compartment [m2]}

\item{from.VOLUME}{Volume of cloudwater compartment [m3]}
}
\description{
Calculation of the first order rate constant for particle deposition from cloudwater compartment to the soil or water surface [s-1]
}
