% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kacompw.R
\name{Kacompw air water partitioning coefficient}
\alias{Kacompw air water partitioning coefficient}
\title{Dimensionless air/water PARTITION COEFFICIENT of molecular species specific to air compartment (scale)}
\arguments{
\item{Pvap25}{vapour pressure of original species at 25°C [Pa]}

\item{MaxPvap}{maximum vapour pressure of original species at 25°C [Pa]}

\item{ChemClass}{see QSAR table}

\item{Sol25}{water solubility of original species at 25°C [mol·m-3]}

\item{T25}{const (25°C ==) 298 K}

\item{Tm}{melting temp}
}
\value{
Kacompw
}
\description{
assess the Kaw if it's not given, if Pvap25 is NA or too large, MaxPvap is used
}
