% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kacompw.R
\name{Kacompw}
\alias{Kacompw}
\title{Dimensionless air/water partitioning coefficient of molecular species specific to air compartment (scale)}
\usage{
Kacompw(
  Kaw25,
  ChemClass,
  Pvap25,
  Sol25,
  H0sol,
  MaxPvap,
  T25,
  Tm,
  Tm_default,
  Temp,
  MW,
  Test
)
}
\arguments{
\item{ChemClass}{chemical class, see QSAR table}

\item{Pvap25}{vapor pressure of original species at 25 C [Pa]}

\item{Sol25}{water solubility of original species at 25 C [mol.m-3]}

\item{H0sol}{Enthalpy of dissolution [-]}

\item{MaxPvap}{maximum vapor pressure of original species at 25 C [Pa]}

\item{T25}{const (25 C ==) 298 K [K]}

\item{Tm}{melting temp [K]}

\item{Tm_default}{default melting point if melting temperature is not given [K]}

\item{Temp}{temperature of compartment [K]}

\item{MW}{molar weight[kg mol-1]}

\item{Test}{determines if SB4-Excel approach is taken or enhanced method from R version [boolean]}
}
\value{
Kacompw
}
\description{
assess the Kaw if it's not given, if Pvap25 is NA or too large, MaxPvap is used
}
