% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kaers.R
\name{Kaers}
\alias{Kaers}
\title{Aerosol solids - air Partition coefficient}
\usage{
Kaers(
  Kaw25,
  Kow,
  Corg,
  RhoCOL,
  Matrix,
  Pvap25,
  MaxPvap,
  Sol25,
  MW,
  T25,
  ChemClass
)
}
\value{
Kaers
}
\description{
Fraction of some fraction
}
