% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kaerw.R
\name{Kaerw}
\alias{Kaerw}
\title{Aerosol water - air Partition coefficient}
\usage{
Kaerw(Kacompw, FRorig, SubCompartName)
}
\arguments{
\item{Kacompw}{air water partitioning coefficient [-]}

\item{FRorig}{fraction of original species [-]}

\item{SubCompartName}{subcompartment considered}
}
\value{
Kaerw
}
\description{
pm
}
