% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_kdegcalc.R
\name{kdegcalc}
\alias{kdegcalc}
\title{Estimate of degradation rate constant}
\usage{
kdegcalc(
  Q.10,
  Ksw,
  Biodeg,
  C.OHrad.n,
  Ea.OHrad,
  k0.OHrad,
  CorgStandard,
  RHOsolid,
  Matrix
)
}
\arguments{
\item{Ksw}{}

\item{Biodeg}{Biodegradability test result (r / r- / i / p)}
}
\description{
If the degradation rate in sediment is not mentioned in the substance
 data, it can be obtained from (European Commission, 2003a). Following section 3.3.12 of report 2015-0161.
}
