% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_KocAltDorC.R
\name{KocAltDorC}
\alias{KocAltDorC}
\title{Input or calculated organic carbon partitioning coefficient for the alternative form.}
\usage{
KocAltDorC(Kow, a, b, pKa, KocAlt, ChemClass)
}
\arguments{
\item{Kow}{Octanol water partitioning coefficient in data}

\item{a}{see QSAR table}

\item{b}{see QSAR table}

\item{pKa}{Dissociation constant of (conjugated) acid (default = 7)}

\item{KocAlt}{Organic Carbon partitioning coefficient in data}
}
\description{
Function that either calculates the Organic Carbon partitioning coefficient from established QSARs based on Kow and pKa
Or uses input data for KocAlt.
}
