% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_KocDorC.R
\name{KocDorC}
\alias{KocDorC}
\title{Input or calculated organic carbon partitioning coefficient for the original form..}
\usage{
KocDorC(Kow, a, b, Koc, ChemClass)
}
\arguments{
\item{Kow}{Octanol water partitioning coefficient in data}

\item{a}{see QSAR table}

\item{b}{see QSAR table}

\item{Koc}{Organic Carbon partitioning coefficient in}

\item{pKa}{Dissociation constant of (conjugated) acid (default = 7)}
}
\description{
Function that either calculates the Organic Carbon partitioning coefficient from established QSARs based on Kow and pKa
Or uses input data for Koc.
}
