% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kp.R
\name{Kp}
\alias{Kp}
\title{Kp}
\usage{
Kp(FRorig, Ksw, Ksw.alt, RHOsolid, Corg, CorgStandard, Matrix)
}
\arguments{
\item{FRorig, }{}

\item{Ksw, }{}

\item{Ksw.alt, }{}

\item{Matrix}{the medium, the formula is only applicable to soil and sediment}
}
\description{
subcompartment/water PARTITION COEFFICIENT
}
