% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_KpCOL.R
\name{KpCOL}
\alias{KpCOL}
\title{Kp for water colloids}
\usage{
KpCOL(D, Matrix)
}
\arguments{
\item{D}{octanol/water partitioning coefficient at neutral pH for colloids [-]}

\item{Matrix}{the medium, the formula is only applicable to soil and sediment}
}
\description{
partitioning coefficient for water colloids
}
