% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kscompw.R
\name{Kscompw}
\alias{Kscompw}
\title{Dimensionless soil/water Partition Coefficient of molecular species specific to soil subcompartment}
\usage{
Kscompw(FRACw, FRACa, Kacompw, FRorig_spw, Kp, RHOsolid, Matrix)
}
\arguments{
\item{FRACa}{NB FRACs is calculated as the remiander from FRACa and FRACw}

\item{Kp}{(soil/water PARTITION COEFFICIENT)}
}
\description{
Dimensionless soil/water PARTITION COEFFICIENT
}
