% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Ksdcompw.R
\name{Ksdcompw}
\alias{Ksdcompw}
\title{Dimensionless sed/water PARTITION COEFFICIENT for molecular species specific to sediment compartment}
\usage{
Ksdcompw(FRACw, FRACs, Kp, all.rhoMatrix, Matrix)
}
\arguments{
\item{Kp}{(sed/water PARTITION COEFFICIENT) of sediment}
}
\description{
Dimensionless sed/water PARTITION COEFFICIENT
}
