% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_kSense.R
\name{kSense}
\alias{kSense}
\title{kSense}
\usage{
kSense(ParentModule, knames = NULL, tol = 1e-30)
}
\arguments{
\item{ParentModule}{SBCore}
}
\value{
States (i) (=mass)
}
\description{
sensitivity Mass 4 kaas,
#logged sd / intervals ??
 by changing kaas what is the change in Mass
}
