% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_KswDorC.R
\name{KswDorC}
\alias{KswDorC}
\title{Ksw}
\usage{
KswDorC(KocDorC, CorgStandard, all.rhoMatrix, Ksw, ChemClass)
}
\arguments{
\item{CorgStandard}{Standard mass FRACTION organic carbon in soil/sediment [-]}

\item{Ksw}{soil water partitioning coefficient in data}

\item{ChemClass}{Class of chemical, see QSAR table (REACH, 2012)}

\item{Kow}{octanol water partitioning coefficient [-]}

\item{a}{see QSAR table}

\item{b}{see QSAR table}

\item{rhoMatrix}{density of the matrix [kg/m3]}

\item{pKa}{Dissociation constant of (conjugated) acid (default = 70}
}
\description{
soil-water partitioning coefficient for organic colloids
}
