% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_MasConc_Otherparticle.R
\name{MasConc_Otherparticle}
\alias{MasConc_Otherparticle}
\title{MasConc_Otherparticle}
\usage{
MasConc_Otherparticle(mConcSusp, mConcCol, SpeciesName)
}
\arguments{
\item{mConcSusp}{concentration of ... in}

\item{mConcCol}{concentration of ... in}

\item{SpeciesName}{yes, its name}
}
\value{
MasConc_Otherparticle, mass concentration of "other" particle before
}
\description{
MasConc_Otherparticle
}
