% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Mass2Conc.R
\name{Mass2Conc}
\alias{Mass2Conc}
\title{Conversion factor mass to concentration}
\usage{
Mass2Conc(Volume, SubCompartName, all.rhoMatrix, FRACa, FRACw)
}
\arguments{
\item{Volume}{of the Box (state) [m3]}

\item{all.rhoMatrix}{density of the matrix [kg m-3]}

\item{FRACa}{fraction of air in media/matrix [-]}

\item{FRACw}{fraction of water in media/matrix [-]}

\item{Matrix}{Media or Matrix type}
}
\value{
Mass2Conc factor
}
\description{
Mass2Conc returns the factor to calculate relevant concentrations in reach compartment e.g. to get[kg/kg w] or [kg/m3]
}
