% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_MTC_2s.R
\name{MTC_2s}
\alias{MTC_2s}
\title{partial mass transfer coefficient from air to soil}
\usage{
MTC_2s(Mackay1, Mackay2, Matrix)
}
\arguments{
\item{Mackay1}{constant described by Mackay (2001) [m/s] https://doi.org/10.1201/9781420032543}

\item{Mackay2}{constant described by Mackay (2001) [-] https://doi.org/10.1201/9781420032543}

\item{Matrix}{matrix considered}
}
\value{
MTC_2w
}
\description{
partial mass transfer coefficient from air to soil
}
