% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_MTC_2sd.R
\name{MTC_2sd}
\alias{MTC_2sd}
\title{PARTIAL MASS TRANSFER COEFFICIENT water to sediment}
\usage{
MTC_2sd(kwsd.water, Matrix, SubCompartName)
}
\arguments{
\item{kwsd.water}{Constant water transfer to sediment [m s-1]}

\item{Matrix}{type of compartment}

\item{SubCompartName}{name of subcompartment}
}
\value{
MTC_2sd
}
\description{
partial mass transfer coefficient water to sediment
}
