% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_MTC_2w.R
\name{MTC_2w}
\alias{MTC_2w}
\title{PARTIAL MASS TRANSFER COEFFICIENT air and sediment to water}
\usage{
MTC_2w(WINDspeed, MW, kwsd.sed, Matrix)
}
\arguments{
\item{WINDspeed}{Windspeed in compartment/scale [m.s-1]}

\item{MW}{Molecular weight of compound kg [g.mol-1]}

\item{kwsd.sed}{Constant sediment transfer rate to water [m/s-1]}

\item{from.Matrix}{Matrix/compartment from which the relevant process is taking place}
}
\value{
MTC_2w
}
\description{
mass transfer coefficient to water
}
