% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rho_species.R
\name{rho_species}
\alias{rho_species}
\title{Rho (density) of species}
\usage{
rho_species(
  SpeciesName,
  SubCompartName,
  ScaleName,
  RhoS,
  RadS,
  RadCOL,
  RadCP,
  RhoCOL,
  RhoCP,
  RhoNuc,
  RadNuc,
  NumConcNuc,
  NumConcAcc,
  Shortest_side,
  Intermediate_side,
  Longest_side,
  Shape,
  Regional_and_Continental_deepocean
)
}
\arguments{
\item{SpeciesName}{Dimension}

\item{SubCompartName}{Dimension}

\item{RhoS}{nanoparticle density [kg/m3]}

\item{RadS}{nanoparticle radius [m]}

\item{RhoNuc}{Nucleation mode aerosol particle  [kg/m3]}

\item{RadNuc}{Nucleation mode aerosol particle radius [m]}

\item{NumConcNuc}{Number concentration of Nucleation mode aerosol particles [#/m3]}

\item{NumConcAcc}{Number concentration of Accumulation mode aerosol particles [#/m3]}

\item{Regional_and_Continental_deepocean}{If this variable is TRUE, Regional and Continental deepocean compartments are removed}

\item{Df}{fractal dimension of combined heteroagglomerate [-]}

\item{RadAcc}{Accumulation mode aerosol particle radius [m]}

\item{RhoAcc}{Accumulation mode aerosol particle  [kg/m3]}
}
\value{
rho_species, approach to calculate density of species in air or water [kg.m-3]
}
\description{
Rho (density) of species
}
