% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_RunoffFlow.R
\name{Runoff}
\alias{Runoff}
\alias{RunoffFlow}
\title{Runoff flow from soil to water}
\usage{
RunoffFlow(FRACrun, Area, RAINrate, Compartment)
}
\arguments{
\item{FRACrun}{volume fraction of precipitation on regional and 
continental natural, agricultural and other soil run off to surface water [-]}

\item{Area}{area of the subcompartment within the scale}

\item{RAINrate}{Precipitation for each subcompartment [m/s]}

\item{Compartment}{Only runoff for Compartment == "soil"}
}
\value{
Rain water runoff from soils [m3/s]
}
\description{
The runoff is from soil to water. Normally this runs into a river, but, 
as there are no rivers modeled in the global scales, it is directed to the sea immediately there. 
Basically, it's a flow, but for this reason, modeled as a variable.
}
