% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_SB1Solve.R
\name{SB1Solve}
\alias{SB1Solve}
\title{SB1Solve}
\usage{
SB1Solve(ParentModule, tol = 1e-30)
}
\arguments{
\item{ParentModule}{SBcore}

\item{tol}{tolerance for accepting as steady state}
}
\value{
States (i) (=mass)
export
}
\description{
solve system of 1rst order k(i,j) and emissions,
 by solving v = 0
}
