% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_SBsolve.R
\name{SBsolve}
\alias{SBsolve}
\title{SBsolve}
\usage{
SBsolve(ParentModule, tmax = 1e+10, nTIMES = 100, EmisAsPulse = F)
}
\arguments{
\item{ParentModule}{SBcore object; to access SimpleBoxODE, states etc. 
too much and to diverse to put them in parameters and complicate the execute method}

\item{tmax}{time [s] for the simulation period}

\item{nTIMES}{number of timesteps}

\item{EmisAsPulse}{T is a pulse-type emission, false is not}
}
\value{
see: ode()
}
\description{
solve system of States for 1rst order k(i,j) and emissions
}
