% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6SBstates.R
\name{SBstates}
\alias{SBstates}
\title{states}
\description{
object in both SBcore as in SolverModule (slimmed down to existing kaas).
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{asDataFrame}}{convienent returning states in a data.frame}

\item{\code{nStates}}{just 4 convenience}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SBstates-new}{\code{SBstates$new()}}
\item \href{#method-SBstates-matchi}{\code{SBstates$matchi()}}
\item \href{#method-SBstates-findState}{\code{SBstates$findState()}}
\item \href{#method-SBstates-clone}{\code{SBstates$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-new"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-new}{}}}
\subsection{Method \code{new()}}{
init, based on
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$new(StatesAsDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{StatesAsDataFrame}}{dataframe of states with columns The3D Abbr and counter i}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-matchi"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-matchi}{}}}
\subsection{Method \code{matchi()}}{
return match (vector of i) in this states. i is the very original order of states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$matchi(vi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vi}}{with i in order of vi}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-findState"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-findState}{}}}
\subsection{Method \code{findState()}}{
map vector abbr for scale species subcompart to index in State
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$findState(abbr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{abbr}}{Oldschool abbreviation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of indices
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
