% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_SBsteady.R
\name{SBsteady}
\alias{SBsteady}
\title{SBsteady}
\usage{
SBsteady(ParentModule, tmax = 1e+10)
}
\arguments{
\item{ParentModule}{passing the SolverModule for access to it's public functions}

\item{tmax}{}
}
\value{
see: runsteady()
export
}
\description{
run system of 1rst order k(i,j) and emissions,
 till steady state
}
