% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_SettlingVelocity.R
\name{SettlingVelocity}
\alias{SettlingVelocity}
\alias{f_SettlingVelocity}
\title{SettlingVelocity}
\usage{
f_SettlingVelocity(rad_species, rho_species, matrix.Rho, DynVisc, Matrix)
}
\arguments{
\item{rad_species}{Radius of the particle [m]}

\item{rho_species}{Density of particle [kg/m3]}

\item{matrix.Rho}{Density of fluid matrix in which particle is present [kg/m3]}

\item{DynVisc}{Dynamic viscosity of the fluid matrix [unit]}

\item{Matrix}{function is defined for Water and Air; slightly different algorithm [text]}
}
\value{
Settling velocity
}
\description{
Settling Velocity or terminal Velocity of particles in a fluid medium based on ...(REF)
NB SettlingVelocity is a function to be used for all types of partical species, not just nano-species. Therefor it's not a variable defining function.
}
