% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6StateModule.R
\name{StateModule}
\alias{StateModule}
\title{State Module}
\description{
Checks if data is consistent with states
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StateModule-new}{\code{StateModule$new()}}
\item \href{#method-StateModule-CheckDataStates}{\code{StateModule$CheckDataStates()}}
\item \href{#method-StateModule-clone}{\code{StateModule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-StateModule-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateModule$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateModule-CheckDataStates"></a>}}
\if{latex}{\out{\hypertarget{method-StateModule-CheckDataStates}{}}}
\subsection{Method \code{CheckDataStates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateModule$CheckDataStates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StateModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StateModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StateModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
