% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fGeneral.R
\name{SteadyStateSolver2}
\alias{SteadyStateSolver2}
\title{Steady state solver function}
\usage{
SteadyStateSolver2(k, e, parms)
}
\arguments{
\item{k}{the first order rate constant matrix [s-1]}

\item{parms}{= empty list (needed so that the SteadyStateSolver and DynamicSolver can be called in the same manner from the SolverModule)}

\item{m}{emission vector [kg.s-1]}
}
\value{
dm (i) = change in mass as list
}
\description{
Using the SimpleBoxODE function to solve the k matrix with emission (e) for steady state.
This uses the steady function from the rootSolve package with possitive = TRUE.
}
