% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Tempfactor.R
\name{Tempfactor}
\alias{Tempfactor}
\title{Temperature correction degradation rate water/sed/soil}
\usage{
Tempfactor(Q.10, Temp, T25, Ea.OHrad, Matrix, SpeciesName)
}
\arguments{
\item{Q.10}{Degradation rate increase factor per  10 C [-]}

\item{Temp}{Average temeperature [K]}

\item{T25}{25 degrees C [K]}

\item{Matrix}{compartment type considered}

\item{SpeciesName}{species considered}

\item{EA.OHrad}{activation energy OH radical reaction [J mol-1]}
}
\value{
Temperature correction degradation rate water/sed/soil [-]
}
\description{
Temperature correction degradation rate water/sed/soil
}
