% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Tempfactor.wsds.R
\name{Tempfactor.wsds}
\alias{Tempfactor.wsds}
\title{Temperature correction degradation rate water/sed/soil}
\usage{
Tempfactor.wsds(Q.10, Temp, T25)
}
\arguments{
\item{Q.10}{Degradation rate increase factor per /U+00B0 10 C [-]}

\item{Temp}{Average precipitation [K]}
}
\value{
Temperature correction degradation rate water/sed/soil [-]
}
\description{
Temperature correction degradation rate water/sed/soil
}
