% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_OtherkAir.R
\name{v_OtherkAir}
\alias{v_OtherkAir}
\alias{OtherkAir}
\title{v_OtherkAir}
\usage{
OtherkAir(all.kaas, ScaleName, SpeciesName)
}
\arguments{
\item{all.kaas}{all other transfer rates in air, used in deposition flux}

\item{ScaleName}{scale name considered}

\item{Species}{species type considered (molecular)}
}
\description{
all other processes in air, used in
}
