% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_SettlingVelocity.R
\name{v_SettlingVelocity}
\alias{v_SettlingVelocity}
\alias{SettlingVelocity}
\title{SettlingVelocity}
\usage{
SettlingVelocity(
  rad_species,
  rho_species,
  rhoMatrix,
  DynViscWaterStandard,
  DynViscAirStandard,
  Matrix,
  SubCompartName,
  ScaleName,
  Shape,
  Longest_side,
  Intermediate_side,
  Shortest_side,
  DragMethod,
  MinSettVel,
  Regional_and_Continental_deepocean,
  VelInput
)
}
\arguments{
\item{rad_species}{Radius of the particle [m]}

\item{rho_species}{Density of particle [kg/m3]}

\item{rhoMatrix}{Density of fluid matrix in which particle is present [kg/m3]}

\item{DynViscWaterStandard}{Dynamic viscosity of the fluid matrix [kg m-1 s-1]}

\item{DynViscAirStandard}{Dynamic viscosity of the fluid matrix [kg m-1 s-1]}

\item{Matrix}{function is defined for Water and Air; slightly different algorithm [text]}

\item{SubCompartName}{name of the different subcompartments, used to segregate different formulas [text]}

\item{Shape}{Shape as defined by user, different possibilities (see f_Vol for options) [text]}

\item{Longest_side}{the longest side of the particle as defined by the user [m]}

\item{Intermediate_side}{the intermediate side of the particle as defined by user [m]}

\item{Shortest_side}{the shortst side of the particle as identified by user [m]}

\item{DragMethod}{The Method used for computing the drag coefficient as defined by user, opportunity for choosing 4 different ones. See f_DragCoefficient for options}

\item{Regional_and_Continental_deepocean}{If this variable is TRUE, Regional and Continental deepocean compartments are removed}
}
\value{
Settling velocity [m.s-1]
}
\description{
Settling Velocity or terminal Velocity of particles in a fluid medium based on different DragMethods.
NB SettlingVelocity is a function to be used for all types of partical species, not just nano-species. Therefor it's not a variable defining function.
}
