% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6VariableModule.R
\name{VariableModule}
\alias{VariableModule}
\title{Variable Module}
\description{
Module dealing with computing and initializing the variables
}
\section{Super class}{
\code{\link[sboo:CalcGraphModule]{sboo::CalcGraphModule}} -> \code{VariableModule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VariableModule-execute}{\code{VariableModule$execute()}}
\item \href{#method-VariableModule-new}{\code{VariableModule$new()}}
\item \href{#method-VariableModule-clone}{\code{VariableModule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableModule-execute"></a>}}
\if{latex}{\out{\hypertarget{method-VariableModule-execute}{}}}
\subsection{Method \code{execute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableModule$execute(debugAt = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-VariableModule-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableModule$new(TheCore, exeFunction, AggrBy, AggrFun)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariableModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VariableModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
