% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Volume.R
\name{Volume}
\alias{Volume}
\title{Volume}
\usage{
Volume(VertDistance, Area, FRACcldw, SubCompartName)
}
\arguments{
\item{VertDistance}{hight or depth for subcompartments below the horizon}

\item{Area}{in m2}

\item{FRACcldw}{fraction of cloudwater}
}
\value{
Volume
}
\description{
Volume
}
