% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_vUncertain.R
\name{vUncertain}
\alias{vUncertain}
\title{vUncertain}
\usage{
vUncertain(
  ParentModule,
  vnamesDistSD,
  n,
  TargetScale = "Regional",
  TargetSubCompart = "river",
  Targetspecies = "Molecular",
  tol = 1e-30
)
}
\arguments{
\item{ParentModule}{SBCore}

\item{vnamesDistSD}{dataframe with columns vnames, distNames (see lhs package for possible distributions), secondPar}

\item{n}{samplesize}
}
\value{
States (i) (=mass)
}
\description{
sensitivity Mass 4 kaas,
#logged sd / intervals ??
 by changing kaas what is the change in Mass
}
