% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ContSea2Moder.R
\name{x_ContSea2Moder}
\alias{x_ContSea2Moder}
\title{ContSea2Moder}
\usage{
x_ContSea2Moder(Volume, TAUsea, all.x_RegSea2Cont, ScaleName, SubCompartName)
}
\arguments{
\item{Volume}{in m3}

\item{TAUsea}{average residence time}

\item{x_RegSea2Cont}{flux from regional sea to continental sea}
}
\value{
water flux; circulation (V/Tau) - sea flux Regional 2 Continental (compensating flux originating from rainwater ??)
}
\description{
ContSea2Moder
}
