% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ContSea2Reg.R
\name{x_ContSea2Reg}
\alias{x_ContSea2Reg}
\title{ContSea2Reg}
\usage{
x_ContSea2Reg(all.x_RiverDischarge)
}
\arguments{
\item{RiverDischarge}{river discharge rate, see x_RiverDischarge [s-1]}
}
\value{
additional flux from continental sea to regional
}
\description{
THe mixing of continental and regional sea's  is based on a factor of 9 times the discharge rate of the regional freshwater to regional sea.
}
