% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_FromModerate2ContWater.R
\name{x_FromModerate2ContWater}
\alias{x_FromModerate2ContWater}
\title{Mixing of upper and deeper sea water layers (w2 and w3)}
\usage{
x_FromModerate2ContWater(
  all.Volume,
  all.TAUsea,
  all.x_RegSea2Cont,
  SubCompartName,
  ScaleName
)
}
\arguments{
\item{SubCompartName}{name of the subcompartment of the box at hand}

\item{ScaleName}{name of the scale of the box at hand}

\item{OceanCurrent}{[m3.s-1]}
}
\value{
Advection sea - deepocean
}
\description{
Mixing of upper and deeper sea water layers (w2 and w3)
}
