% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_LakeOut.R
\name{x_LakeOut}
\alias{x_LakeOut}
\title{LakeOut}
\usage{
x_LakeOut(
  RainOnFreshwater,
  all.RunoffFlow,
  FracROWatComp,
  SubCompartName,
  ScaleName
)
}
\arguments{
\item{RainOnFreshwater}{The direct rain on the freshwater compartment (e.g. a lake).}

\item{all.RunoffFlow}{RunoffFlow from soil to the freshwater compartment}

\item{FracROWatComp}{The fraction runoff discharging into the separate w0 freshwater compartment based on surface area fraction.}

\item{SubCompartName}{Only applies to lake compartment, so the compartment name is needed.}

\item{ScaleName}{Runoff needs to be calculated per scale.}
}
\value{
Lake discharge
}
\description{
This calculates the flow from a freshwater (w0) compartment (e.g. lake) within the regional or continental scale to the main freshwater compartment (w1).
}
