% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_LakeOutflow.R
\name{x_LakeOutflow}
\alias{x_LakeOutflow}
\title{LakeOutflow}
\usage{
x_LakeOutflow(
  all.x_RiverDischarge,
  all.x_ContRiver2Reg,
  LakeFracRiver,
  ScaleName
)
}
\arguments{
\item{LakeFracRiver}{Fraction (default 0.1) of r}

\item{ScaleName}{different formula for scale Continental or Regional}

\item{RiverDischarge}{see ppt}

\item{ContRiver2Reg}{see ppt}
}
\value{
River Discharge
}
\description{
related to what flows from the river into the sea
}
