% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_OceanMixing2Deep.R
\name{x_OceanMixing2Deep}
\alias{x_OceanMixing2Deep}
\title{Mixing of upper and deeper sea water layers (w2 and w3)}
\usage{
x_OceanMixing2Deep(Volume, TAUsea, OceanCurrent, ScaleName, SubCompartName)
}
\arguments{
\item{OceanCurrent}{[m3.s-1]}

\item{ScaleName}{name of the scale of the box at hand}

\item{to.Volume}{Volume of the to box}

\item{from.Volume}{Volume of the from box}

\item{to.TAUsea}{TAU - Residence time in sea; scale variable}

\item{from.TAUsea}{TAU - Residence time in sea; for the "to" box}
}
\value{
Advection sea - deepocean
}
\description{
Mixing of upper and deeper sea water layers (w2 and w3)
}
