% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_OceanMixing2Sea.R
\name{x_OceanMixing2Sea}
\alias{x_OceanMixing2Sea}
\title{Mixing of deep ocean to surface sea water layer (w2 and w3)}
\usage{
x_OceanMixing2Sea(all.Volume, TAUsea, OceanCurrent, SubCompartName, ScaleName)
}
\arguments{
\item{OceanCurrent}{[m3.s-1]}

\item{SubCompartName}{name of the subcompartment of the box at hand}

\item{ScaleName}{name of the scale of the box at hand}

\item{to.Volume}{Volume of the to box}

\item{from.Volume}{Volume of the from box}

\item{to.TAUsea}{TAU - Residence time in sea; scale variable}

\item{from.TAUsea}{TAU - Residence time in sea; for the "to" box}
}
\value{
Mixing flow from deep ocean to surface sea [m3.s-1]
}
\description{
Mixing of deep ocean to surface sea water layer (w2 and w3)
}
