% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_RegSea2Cont.R
\name{x_RegSea2Cont}
\alias{x_RegSea2Cont}
\title{RegSea2Cont}
\usage{
x_RegSea2Cont(all.x_RiverDischarge, all.x_ContSea2Reg)
}
\arguments{
\item{RiverDischarge}{find the flux in all. RiverDischarge fluxes}

\item{ContSea2Reg}{find the ContSea2Reg flux in all.(ContSea2Reg)fluxes (there is no direct relation)}
}
\value{
River Discharge for scale Continental
}
\description{
RegSea2Cont
}
