% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_RiverDischarge.R
\name{x_RiverDischarge}
\alias{x_RiverDischarge}
\title{RiverDischarge [s-1]}
\usage{
x_RiverDischarge(
  all.RunoffFlow,
  all.RainOnFreshwater,
  dischargeFRAC,
  all.x_ContRiver2Reg,
  ScaleName,
  SubCompartName
)
}
\arguments{
\item{dischargeFRAC}{Fraction discharge of freshwater between regional and continental scales and vice versa [-]}

\item{ScaleName}{Name of the scale of the box at hand}

\item{SubCompartName}{Name of the subcompartment of the box at hand}

\item{RunoffFlow}{RunoffFlow flow from soil to river [m3.s-1]}

\item{RainOnFreshwater}{Water flow of rain directly on lake/river [m3.s-1]}

\item{x_ContRiver2Reg}{Flow from continental to regional river water [s-1]}
}
\value{
River Discharge [s-1]
}
\description{
RiverDischarge [s-1]
}
