% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_RiverSeaScales.R
\name{x_RiverSeaScales}
\alias{x_RiverSeaScales}
\title{x_RiverSeaScales}
\usage{
x_RiverSeaScales(
  from.SubCompartName,
  from.ScaleName,
  to.ScaleName,
  ContRiver2Reg,
  ContSea2Reg,
  RegSea2Cont,
  OceanCurrent,
  from.TAUsea,
  to.TAUsea,
  ContinentalInModerate
)
}
\arguments{
\item{from.SubCompartName}{dimension dependency in function}

\item{from.ScaleName}{dimension dependency in function}

\item{to.ScaleName}{dimension dependency in function}

\item{ContRiver2Reg}{riverflow from continental to regional}

\item{ContSea2Reg}{seaflow from continental to regional}

\item{RegSea2Cont}{seaflow from regional to continental}

\item{OceanCurrent}{global constant}

\item{from.TAUsea}{refresh rate}

\item{to.TAUsea}{dito for the destination box}

\item{ContinentalInModerate}{logical; continental can be in Moderate or in Tropic}

\item{from.Volume}{in m3}

\item{to.Volume}{in m3}
}
\value{
Advection seas and deepoceans, excluding vertical mixing sea <-> deepocean
}
\description{
x_RiverSeaScales
}
