% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_AirFlow.R
\name{AirFlow}
\alias{AirFlow}
\title{AirFlow}
\usage{
AirFlow(Volume, Area, WINDspeed, SubCompartName)
}
\arguments{
\item{Volume}{in m3}

\item{Area}{in m2}

\item{WINDspeed}{m.s-1}

\item{SubCompartName}{for which the calculation is executed}
}
\value{
AirFlow, from one scale to another
}
\description{
AirFlow
}
