% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Area.R
\name{Area}
\alias{Area}
\title{Area}
\usage{
Area(AreaLand, AreaSea, landFRAC, all.landFRAC, SubCompartName, ScaleName)
}
\arguments{
\item{landFRAC}{subcompartment fraction of AREAland}

\item{SubCompartName}{for which Area is calculated}

\item{ScaleName}{for which Area is calculated}

\item{AREAland}{AREA of land in scale}

\item{AREAsea}{AREA of sea in scale}
}
\value{
Area (but not for sediment) based on data for the SubCompartment / Scale
}
\description{
Area
}
