% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_AreaLand.R
\name{AreaLand}
\alias{AreaLand}
\title{AreaLand}
\usage{
AreaLand(all.TotalArea, all.FRACsea, ScaleName)
}
\arguments{
\item{all.TotalArea}{total AREA of Scale, possibly including other scales,
hence the all. (nesting!)}

\item{all.FRACsea}{sea fraction of scale, possibly including other scales,
hence the all. (nesting!)}

\item{ScaleName}{indicating for which scale the function is called}
}
\value{
Land in SystemArea
}
\description{
very similar to AreaSea; Keep It Stupid Simple...
}
