% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_AreaSea.R
\name{AreaSea}
\alias{AreaSea}
\title{AreaSea}
\usage{
AreaSea(all.TotalArea, all.FRACsea, ScaleName)
}
\arguments{
\item{all.TotalArea}{total AREA of Scale, possibly including other scales
by nesting, hence the all.}

\item{all.FRACsea}{sea fraction of scale, possibly including other scales
by nesting, hence the all.}

\item{ScaleName}{name of the scale of the box}
}
\value{
Area of Sea for scale with name ScaleName
}
\description{
AreaSea
}
