% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSClassicNano.R
\name{ClassicNanoWorld}
\alias{ClassicNanoWorld}
\title{ClassicNanoWorld}
\description{
the centre object for running simple box with an R script.
}
\section{Super class}{
\code{\link[sboo:StateModule]{sboo::StateModule}} -> \code{ClassicNanoWorld}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{varOrigine}}{getter for r.o. property
#description extend subcompart with compart properties}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClassicNanoWorld-new}{\code{ClassicNanoWorld$new()}}
\item \href{#method-ClassicNanoWorld-clone}{\code{ClassicNanoWorld$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sboo" data-topic="StateModule" data-id="CheckDataStates"><a href='../../sboo/html/StateModule.html#method-StateModule-CheckDataStates'><code>sboo::StateModule$CheckDataStates()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoWorld-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoWorld-new}{}}}
\subsection{Method \code{new()}}{
init
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoWorld$new(MlikeFile, Substance = "default substance")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{MlikeFile}}{location for standard input data}

\item{\code{Substance}}{The substance for which all calculation are.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClassicNanoWorld-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClassicNanoWorld-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClassicNanoWorld$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
