% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Brown.R
\name{f_Brown}
\alias{f_Brown}
\title{Brownian motion or kinetic collision frequency coefficient (20181102)}
\usage{
f_Brown(Temp, viscosity, from.radius, radOther)
}
\arguments{
\item{Temp}{temperature [K]}

\item{viscosity}{Dynamic viscosity of liquid  (fraction of) compartment [kg.m-1.s-1]}

\item{from.radius}{Radius of nanoparticle [m]}

\item{radOther}{Radius of Other particle [m]}
}
\value{
fBrown [s-1]
}
\description{
Collission frequency of ENPs with other particulates due to bownian motion (kinetic energy) in s-1 for heteroagglomeration
}
