% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_CORRsoil.R
\name{f_CORRsoil}
\alias{f_CORRsoil}
\title{Correction factor depth-dependency soil concentrations}
\usage{
f_CORRsoil(VertDistance, relevant_depth_s, penetration_depth_s)
}
\arguments{
\item{relevant_depth_s}{(get discription from Hollander et al. (2004))}

\item{penetration_depth_s}{(get discription from Hollander et al. (2004))}

\item{vertDistance}{depth of regional, continental and global soils [m]}
}
\value{
Correction factor depth dependent soil concentration
}
\description{
Correction factor depth dependent soil concentrations. 
This is used for correcting the rate of runoff and volatilisation 
due to the non-homogeneaus distribution across the top soil layer.
The default relevant depth is 0 cm and the soil penetration depth is 10 cm.
The approach to derive this correction factor can be obtained from Hollander et al. (2004).
}
