% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_DragCoefficient.R
\name{f_DragCoefficient}
\alias{f_DragCoefficient}
\title{Drag Coefficient particles}
\usage{
f_DragCoefficient(DragMethod, Re, Psi, CSF)
}
\arguments{
\item{DragMethod}{Method of calculating the Drag Coefficient}

\item{Re}{Reynolds number, as returned by the solver [-]}

\item{Psi}{Shape factor, circularity/sphericity [-]}

\item{CSF}{Corey Shape Factor [-]}

\item{CD}{Drag Coefficient of a particle [-]}

\item{Dioguardi}{A Drag Coefficient method as described in Dioguardi & Mele (2018) https://doi.org/10.1016/j.powtec.2015.02.062}

\item{Swamee}{A Drag Coefficient method as described in Swamee & Ojha (1991) https://doi.org/10.1061/(ASCE)0733-9429(1991)117:5(660)}

\item{Stokes}{A Stokes-Dietrich approximation as described in Dietrich (1982) https://doi.org/10.1029/WR018i006p01615}
}
\value{
f_DragCoefficient
}
\description{
Computes the Drag Coefficoent based on the DragMethod as defined by user
}
