% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_SchmidtNumber.R
\name{f_SchmidtNumber}
\alias{f_SchmidtNumber}
\title{SchmidtNumber}
\usage{
f_SchmidtNumber(
  visc,
  rhoMatrix,
  Diffusivity = NULL,
  Temp = NULL,
  rad_particle = NULL,
  Cunningham = NULL
)
}
\arguments{
\item{visc}{Dynamic viscosity of air [kg.m-1.s-1]}

\item{rhoMatrix}{Density of the compartment the particle is in [kg m-3]}

\item{Temp}{Temperature of the compartment the particle is in [kg m-3]}

\item{rad_particle}{radius of the particle [m]}
}
\value{
SchmidtNumber
}
\description{
Calculation of the dimensionless Schmidt number, based on Wang et al. (2010) https://doi.org/10.5194/acp-10-5685-2010
Denotes the ratio of momentum diffusivity and mass diffusivity.
}
