% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Cunningham.R
\name{fCunningham}
\alias{fCunningham}
\alias{f_Cunningham}
\title{Cunningham}
\usage{
f_Cunningham(rad_species)
}
\arguments{
\item{rad_species}{radius of the species}
}
\value{
Cunningham
}
\description{
Collission frequency of ENPs with other particulates due to Gravitational or Intertial impaction in s-1 for heteroagglomeration
}
