% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FracWatComp.R
\name{FracROWatComp}
\alias{FracROWatComp}
\title{FracRoWatComp}
\usage{
FracROWatComp(all.landFRAC, all.Matrix, Matrix, SubCompartName, ScaleName)
}
\arguments{
\item{all.landFRAC}{Fractions of land compartments #[-]}

\item{all.Matrix}{Matrix compartment the subcompartments belong to #[-]}

\item{Matrix}{Current matrix the function calculates for [-]}

\item{SubCompartName}{Subcompartment the function calculates for [-]}

\item{ScaleName}{The scale the function calculates for [-]}
}
\value{
Fraction of water component of .to subcompartment #[-]
}
\description{
fraction of water component compared to total water in scale for correction of runoff
}
