% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_FRinaers.R
\name{FRinaers}
\alias{FRinaers}
\title{Fraction of chemical in aerosol solids}
\usage{
FRinaers(Kaers, Kaerw, FRACs, FRACw)
}
\arguments{
\item{FRACaer}{Fraction aerosols in air [-]}
}
\value{
The fraction of a chemical in the aerosol solid phase. Total: FRingas + FRinaerw + FRinaers = 1.
}
\description{
Calculates the fraction of a chemical in aerosol solids relative to the total amount of the chemical in air.
}
\seealso{
[Fringas(), FRinw(), FRins()]
}
