% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_Vol.R
\name{fVol}
\alias{fVol}
\title{Calculate volume of sphere assuming (20181102)}
\usage{
fVol(rad_particle)
}
\arguments{
\item{rad_particle}{Radius of particle [m]}
}
\value{
fVol [m3]
}
\description{
Calculate the volume of spherical particle using the radius in m3
}
