% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Adsorption.R
\name{k_Adsorption}
\alias{k_Adsorption}
\title{Adsorption of chemical to across an interface}
\usage{
k_Adsorption(
  FRingas,
  FRinw,
  from.MTC_2sd,
  to.FRorig_spw,
  to.MTC_2w,
  from.MTC_2w,
  to.MTC_2a,
  from.MTC_2s,
  FRorig,
  Kacompw,
  to.Kscompw,
  to.Matrix,
  VertDistance,
  AreaLand,
  AreaSea,
  Area,
  from.SubCompartName,
  to.SubCompartName,
  ScaleName
)
}
\arguments{
\item{FRingas}{see FRingas()}

\item{FRinw}{see FRinw()}

\item{FRorig}{SBvariable}

\item{Kacompw}{SBvariable}

\item{VertDistance}{data}

\item{MTC_2sd}{SBvariable, see MTC_2sd()}

\item{MTC_2w}{SBvariable}

\item{MTC_2a}{SBvariable}

\item{MTC_2s}{SBvariable}

\item{Matrix}{# can use paraminherit to use description from another function! No need to copy paste same descriptions!}
}
\value{
The adsorption rate constant relevant for the receiving compartments soil, water or sediment [s-1]
}
\description{
The adsorption rate constant (k_Adsorption) is calculated based on the ... See 2015-0161 report.
}
