% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Advection.R
\name{k_Advection}
\alias{k_Advection}
\title{General Advection process}
\usage{
k_Advection(flow, Volume)
}
\arguments{
\item{flow}{advection air}

\item{Volume}{in m3}
}
\value{
Rate constant for 1rst order process associated with fluxes
}
\description{
Calculation of k, given a Flow
}
