% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Advection.R
\name{k_Advection}
\alias{k_Advection}
\title{General Advection process}
\usage{
k_Advection(flow, Volume)
}
\arguments{
\item{flow}{advection rate [m/s]}

\item{Volume}{volume of compartment [m3]}
}
\value{
Rate constant for 1st order process associated with fluxes
}
\description{
Calculation of k, given a Flow
}
