% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_CWscavenging.R
\name{k_CWscavenging}
\alias{k_CWscavenging}
\title{Cloud water scavenging coefficients as function of particle size and density}
\arguments{
\item{rho_species}{particle density []}

\item{rad_species}{particle radius []}

\item{to.rhoMatrix}{density of cloud water []}

\item{rhoMatrix}{density of air []}

\item{DynViscAirStandard}{dynamic viscosity of air []}

\item{DynViscWaterStandard}{dynamic viscosity of (cloud) water []}

\item{Temp}{tempreature [oC]}
}
\value{
k_CWscavenging
}
\description{
a description follows .. Joris? based on Wang, X.; Zhang, L.;
  Moran, M.D. Uncertainty assessment of current size-resolved
  parameterizations for below-cloud particle scavenging by rain. Atmos. Chem.
  Phys. Discuss. 2010, 10, 2503–2548
}
