% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_CWscavenging.R
\name{k_CWscavenging}
\alias{k_CWscavenging}
\title{Cloud water scavenging coefficients as function of particle size and density}
\usage{
k_CWscavenging(
  RAINrate,
  FRACtwet,
  tdry,
  twet,
  COLLECTeff,
  rad_species,
  rho_species,
  to.rhoMatrix,
  from.rhoMatrix,
  DynViscAirStandard,
  DynViscWaterStandard,
  SettlingVelocity,
  Temp,
  Matrix,
  SpeciesName,
  VertDistance,
  SubCompartName
)
}
\arguments{
\item{rad_species}{particle radius []}

\item{rho_species}{particle density []}

\item{to.rhoMatrix}{density of cloud water []}

\item{DynViscAirStandard}{dynamic viscosity of air []}

\item{DynViscWaterStandard}{dynamic viscosity of (cloud) water []}

\item{Temp}{tempreature [oC]}

\item{rhoMatrix}{density of air []}
}
\value{
k_CWscavenging
}
\description{
a description follows .. Joris? based on Wang, X.; Zhang, L.;
  Moran, M.D. Uncertainty assessment of current size-resolved
  parameterizations for below-cloud particle scavenging by rain. Atmos. Chem.
  Phys. Discuss. 2010, 10, 2503–2548
}
