% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Degradation.R
\name{k_Degradation}
\alias{k_Degradation}
\title{k_Degradation}
\usage{
k_Degradation(
  FRingas,
  KdegDorC,
  C.OHrad.n,
  C.OHrad,
  Tempfactor,
  FRinw,
  BACTtest,
  BACTcomp,
  Matrix,
  SpeciesName
)
}
\arguments{
\item{FRingas}{see}

\item{C.OHrad.n}{?}

\item{C.OHrad}{?}

\item{kdeg.air}{kdeg.water kdeg.water}

\item{k0.OHrad}{?}

\item{Ea.OHrad}{?}

\item{Temp}{[K]}

\item{T25}{= 298 K}
}
\value{
Degradation rate constant for molecular species
}
\description{
calculate k for degradation;
}
