% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Degradation.R
\name{k_Degradation}
\alias{k_Degradation}
\title{k_Degradation}
\usage{
k_Degradation(
  FRingas,
  KdegDorC,
  C.OHrad.n,
  C.OHrad,
  Tempfactor,
  FRinw,
  BACTtest,
  BACTcomp,
  Matrix,
  SpeciesName,
  SubCompartName,
  ScaleName,
  Test,
  kdis = NA
)
}
\arguments{
\item{FRingas}{fraction of original species in gas phase, see FRingas [-1]}

\item{KdegDorC}{calculated degradation rate constant, see KdegDorc [s-1]}

\item{C.OHrad.n}{general OH radical concentration, based on Wania & Daly (2002)}

\item{C.OHrad}{OH radical concentration specific to compartment, based on Wania & Daly (2002)}

\item{Tempfactor}{Temperature correction rate air, constant}

\item{BACTtest}{concentration bacteria in test water [CFU mL -1]}

\item{BACTcomp}{concentration bacteria in considered compartment [CFU mL -1]}

\item{Matrix}{compartment type}

\item{SpeciesName}{name of the considered species}

\item{SubCompartName}{name of the subcompartment}

\item{ScaleName}{name of the considered scale}
}
\value{
Degradation rate constant for molecular species
}
\description{
calculate k for degradation, based on observations from Wania & Daly (2002) https://doi.org/10.1016/S1352-2310(02)00693-3
}
