% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Desorption.R
\name{k_Desorption}
\alias{k_Desorption}
\title{Desorbtion of molecular species from sediment to water}
\usage{
k_Desorption(
  Ksdcompw,
  MTC_2w,
  to.MTC_2sd,
  VertDistance,
  SpeciesName,
  to.SubCompartName,
  ScaleName
)
}
\value{
Desorption rate constant from sediment to water [s-1]
}
\description{
Desorbtion of molecular species from sediment to water
}
