% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_DryDeposition.R
\name{k_DryDeposition}
\alias{k_DryDeposition}
\title{Dry deposition from air to surface soil or water of particulate species}
\usage{
k_DryDeposition(
  to.Area,
  from.Volume,
  AEROresist,
  to.gamma.surf,
  FricVel,
  DynViscAirStandard,
  rhoMatrix,
  to.ColRad,
  rad_species,
  rho_species,
  Temp,
  to.alpha.surf,
  SettlingVelocity,
  SpeciesName,
  SubCompartName
)
}
\arguments{
\item{to.Area}{Surface area of receiving land or water compartment [m2]}

\item{from.Volume}{Volume of air compartment [m3]}

\item{AEROresist}{aerodynamic resistance (Constant set to 74) [s.m-1]}

\item{DynViscAirStandard}{Dynamic viscosity of air}

\item{rhoMatrix}{Density of air [kg.m-3]
#param from.viscosity Dynamic viscosity of air compartment [kg.m-1.s-1]}

\item{rad_species}{rad_particle}

\item{rho_species}{rho_particle}

\item{Temp}{Temperature [K]}

\item{to.alpha.surf}{depends on the vegetation type, see table A.1 in ref guide LOTEUR v2.0 2016}

\item{ColRad}{Land surface particle collector radius [m] appendix A of LOTEUR.
#param FricVel Friction velocity [m s-1] (19 according to Van Jaarsveld, table 2.2)}

\item{from.SettlingVelocity}{Settlings Velocity of particulate species in air [m.s-1]}
}
\value{
k_drydeposition, the rate constant for 1rst order process: dry deposition from air to soil or water [s-1]
}
\description{
Calculation of the first order rate constant for deposition from air to the soil or water surface [s-1]
Diffusivity is calculated from f_Diffusivity
}
