% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_DryDeposition.R
\name{k_DryDeposition}
\alias{k_DryDeposition}
\title{Dry deposition from air to surface soil or water of particulate species}
\usage{
k_DryDeposition(
  to.Area,
  from.Volume,
  AEROresist,
  to.gamma.surf,
  FricVel,
  DynViscAirStandard,
  rhoMatrix,
  to.ColRad,
  rad_species,
  rho_species,
  Temp,
  to.alpha.surf,
  SettlingVelocity,
  SpeciesName,
  SubCompartName,
  AEROSOLdeprate,
  Test = FALSE,
  to.Matrix,
  RAINrate,
  from.Matrix,
  from.rhoMatrix,
  FRACtwet,
  to.SubCompartName
)
}
\arguments{
\item{to.Area}{Surface area of receiving land or water compartment [m2]}

\item{from.Volume}{Volume of air compartment [m3]}

\item{AEROresist}{aerodynamic resistance (Constant set to 74) [s.m-1]}

\item{FricVel}{Friction velocity [m s-1] (19 according to Van Jaarsveld, table 2.2)}

\item{DynViscAirStandard}{Dynamic viscosity of air}

\item{rhoMatrix}{density of the matrix [kg m3]}

\item{rad_species}{rad_particle [m]}

\item{rho_species}{rho_particle [kg m3]}

\item{Temp}{Temperature [K]}

\item{to.alpha.surf}{depends on the vegetation type, see table A.1 in ref guide LOTEUR v2.0 2016}

\item{SpeciesName}{name of the species considered}

\item{AEROSOLdeprate}{deposition velocity of aerosol particles [m/s]}

\item{Test}{determines if SB4-Excel approach is taken or enhanced method from R version [boolean]}

\item{FRACtwet}{fraction of wet periods [-]}

\item{ColRad}{Land surface particle collector radius [m] appendix A of LOTEUR.}

\item{gamma.surf}{resistance parameter depending on average surface type as taken from the LOTOS-EUROS reference guide table A.1}

\item{SubcompartName}{name of the Subcompartment}

\item{from.SettlingVelocity}{Settlings Velocity of particulate species in air [m.s-1]}

\item{Matrix}{type of compartment [-]}

\item{Cunningham}{collision frequency of ENPs with other particles [-] see f_Cunnningham}

\item{Diffusivity}{see f_Diffusivity}
}
\value{
k_drydeposition, the rate constant for 1rst order process: dry deposition from air to soil or water [s-1]
}
\description{
Calculation of the first order rate constant for deposition from air to the soil or water surface [s-1]
Adjustments for (Test - FALSE, new SB version) are made based on LOTOS-EUROS reference guide v.2.2002 https://www.rivm.nl/lotos-euros
}
