% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Fragmentation.R
\name{k_Fragmentation}
\alias{k_Fragmentation}
\title{Fragmentation and degradation of plastics}
\usage{
k_Fragmentation(kmpdeg, kfrag, SubCompartName, ScaleName)
}
\arguments{
\item{kmpdeg}{degradation rate of plastic in certain subcompartment [s-1]}

\item{kfrag}{fragmentation rate of plastic in certain subcompartment [s-1]}

\item{SubCompartName}{considered subcompartment}

\item{ScaleName}{considered scale}
}
\value{
k_Fragmentation, the combined degradation and fragmentation rate of microplastics [s-1]
}
\description{
Calculation of fragmentation and degradation of plastics, only used in Plastics World
}
