% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Resuspension.R
\name{k_Resuspension}
\alias{k_Resuspension}
\title{Resuspension rate constant of substances in sediment}
\usage{
k_Resuspension(
  VertDistance,
  DynViscWaterStandard,
  to.rhoMatrix,
  to.NETsedrate,
  to.RadCP,
  to.RhoCP,
  from.RhoCP,
  FRACs,
  to.SUSP,
  SpeciesName,
  ScaleName,
  to.SubCompartName,
  from.SubCompartName,
  Test
)
}
\arguments{
\item{VertDistance}{mixed depth sediment compartment #[m]}

\item{DynViscWaterStandard}{Dynamic viscosity of water [kg m-1 s-1]}

\item{FRACs}{Volume fraction solids in sediment #[-]}

\item{SpeciesName}{species considered}

\item{ScaleName}{scale considered}

\item{Test}{determines if SB4-Excel approach is taken or enhanced method from R version [boolean]}

\item{rhoMatrix}{density of the Matrix [kg m-3]}

\item{NETsedrate}{net sedimentation rate, data input [s] (Schoorl et al., 2015)}

\item{RhoCP}{Mineral density of sediment and soil #[kg/m3]}

\item{RadCP}{radius of coarse particulate particles [m]}

\item{SUSP}{mass concentration of suspended matter [kg m-3]}

\item{SubCompartName}{subcompartment considered}

\item{SettlingVelocitySPM}{settling velocity of suspended matter particles}
}
\value{
k_Resuspension Resuspension flow from sediment #[s-1]
}
\description{
Calculation of the resuspension rate. The top layer of sediment is assumed to be well mixed and thus continiously refreshed,
for additional details see Schoorl et al. (2015)
}
