% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Runoff.R
\name{k_Runoff}
\alias{k_Runoff}
\title{Run-off of particles (unbound to be added)}
\usage{
k_Runoff(
  Runoff,
  VertDistance,
  Volume,
  Kscompw,
  relevant_depth_s,
  penetration_depth_s,
  ScaleName,
  to.SubCompartName,
  SpeciesName
)
}
\arguments{
\item{VertDistance}{Mixing depth soil #[m]}

\item{FRACrun}{Fraction run-off #[-]}

\item{RAINrate}{Average precipitation #[m/s]}
}
\value{
k_Run-off Run-off of particles from soil to water #[s-1]
}
\description{
Run-off of particles (unbound to be added)
}
