% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Sedimentation.R
\name{k_Sedimentation}
\alias{k_Sedimentation}
\title{k_Sedimentation}
\usage{
k_Sedimentation(
  FRinw,
  SettlingVelocity,
  DynViscWaterStandard,
  VertDistance,
  from.RhoCP,
  from.RadCP,
  SpeciesName,
  SubCompartName,
  ScaleName
)
}
\arguments{
\item{DynViscWaterStandard}{Dynamic viscosity of the fluid matrix}

\item{VertDistance}{compartment (mixed/mixk_Sedimentationing) DEPTH [m]}

\item{NaturalRho}{Density of natural particle}

\item{rhoMatrix}{Density of matrix in which particle is present}

\item{NaturalRad}{Radius of the natural particle}

\item{CompartName}{function is defined for Water and Air; slightly different function}
}
\value{
k_Sedimentation, the rate constant for sedimentation as first order process
}
\description{
Calculate the rate constant for sedimentation [s-1]
}
