% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Sedimentation.R
\name{k_Sedimentation}
\alias{k_Sedimentation}
\title{k_Sedimentation}
\usage{
k_Sedimentation(
  FRinw,
  SettlingVelocity,
  DynViscWaterStandard,
  VertDistance,
  from.RhoCP,
  from.RadCP,
  RadS,
  SpeciesName,
  SubCompartName,
  to.SubCompartName,
  ScaleName,
  Test
)
}
\arguments{
\item{FRinw}{Fraction chemical dissolved in water [-]}

\item{SettlingVelocity}{Settling velocity of particulate species [m.s-1]}

\item{DynViscWaterStandard}{Dynamic viscosity of the fluid matrix}

\item{VertDistance}{Compartment depth/height [m]}

\item{SpeciesName}{Name of relevant species (Molecular or particulate) for which k_Sedimentation is being calculated}

\item{SubCompartName}{Name of relevant subcompartment for which k_Sedimentation is being calculated}

\item{ScaleName}{Name of relevant scale for which k_Sedimentation is being calculated}

\item{Test}{determines if SB4-Excel approach is taken or enhanced method from R version [boolean]}

\item{RadCP}{Radius of the Coarse natural particle [m]}

\item{RhoCP}{Density of the Coarse natural particle [m]}
}
\value{
k_Sedimentation, the rate constant for sedimentation as first order process
}
\description{
Calculate the rate constant for sedimentation [s-1]
}
