% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_Volatilisation.R
\name{k_Volatilisation}
\alias{k_Volatilisation}
\title{Volatilisation rate constant}
\usage{
k_Volatilisation(
  to.MTC_2w,
  from.MTC_2a,
  to.MTC_2s,
  Kacompw,
  from.FRorig,
  from.FRorig_spw,
  FRinw,
  Kscompw,
  VertDistance,
  SpeciesName,
  Matrix,
  relevant_depth_s,
  penetration_depth_s,
  from.SubCompartName,
  from.ScaleName,
  to.SubCompartName,
  to.ScaleName
)
}
\arguments{
\item{VertDistance}{Mixing depth soil #[m]}

\item{FRACinf}{Fraction infiltration #[-]}

\item{RAINrate}{Average precipitation #[m/s]}
}
\value{
Volatilisation rate constant [s-1]
}
\description{
volatilisation process based on the two-film resistance model (Schwarzenbach et al., 1993) ISBN: 978-1-118-76723-8
}
