% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kacompw.R
\name{Kacompw}
\alias{Kacompw}
\title{Dimensionless air/water PARTITION COEFFICIENT of molecular species specific to air compartment (scale)}
\usage{
Kacompw(
  Kaw25,
  ChemClass,
  Pvap25,
  Sol25,
  H0sol,
  MaxPvap,
  T25,
  Tm,
  Tm_default,
  Temp,
  MW
)
}
\arguments{
\item{ChemClass}{see QSAR table}

\item{Pvap25}{vapour pressure of original species at 25 C [Pa]}

\item{Sol25}{water solubility of original species at 25 C [mol.m-3]}

\item{MaxPvap}{maximum vapour pressure of original species at 25 C [Pa]}

\item{T25}{const (25 C ==) 298 K}

\item{Tm}{melting temp}
}
\value{
Kacompw
}
\description{
assess the Kaw if it's not given, if Pvap25 is NA or too large, MaxPvap is used
}
