% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kp.R
\name{Kp}
\alias{Kp}
\title{Kp}
\usage{
Kp(
  FRorig,
  KswDorC,
  Ksw.alt,
  all.rhoMatrix,
  Corg,
  CorgStandard,
  Matrix,
  ChemClass
)
}
\arguments{
\item{FRorig, }{}

\item{KswDorC, }{}

\item{Matrix}{the medium, the formula is only applicable to soil and sediment}

\item{ChemClass}{Chemical class, e.g. neutral or metal}

\item{KswDorC.alt, }{}
}
\description{
subcompartment/water PARTITION COEFFICIENT
}
