% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kp.R
\name{Kp}
\alias{Kp}
\title{Kp}
\usage{
Kp(
  FRorig,
  KswDorC,
  Ksw.alt,
  all.rhoMatrix,
  Corg,
  CorgStandard,
  Matrix,
  ChemClass
)
}
\arguments{
\item{FRorig, }{fraction of species in its original form [-]}

\item{KswDorC}{calculated soil water partitioning coefficient [-]}

\item{Corg}{mass fraction organic carbon in soil/sediment [-]}

\item{CorgStandard}{standard mass fraction organic carbon in soil/sediment [-]}

\item{Matrix}{the medium, the formula is only applicable to soil and sediment}

\item{ChemClass}{Chemical class, e.g. neutral or metal}

\item{RHOsolid}{density of the solid [kg m-3]}

\item{ksw.alt}{}
}
\description{
subcompartment/water partitioning coefficient
}
