% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_KpCOL.R
\name{KpCOL}
\alias{KpCOL}
\title{Kp for water colloids}
\usage{
KpCOL(D, Matrix)
}
\arguments{
\item{Matrix}{the medium, the formula is only applicable to soil and sediment}
}
\description{
subcompartment/water PARTITION COEFFICIENT
}
