% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Kscompw.R
\name{Kscompw}
\alias{Kscompw}
\title{Dimensionless soil/water Partition Coefficient of molecular species specific to soil subcompartment}
\usage{
Kscompw(FRACw, FRACa, Kacompw, FRorig_spw, Kp, all.rhoMatrix, Matrix)
}
\arguments{
\item{FRACw}{fraction of water in compartment [-]}

\item{FRACa}{fraction of air in compartment [-]}

\item{FRorig_spw}{fraction of original species in soil pore water [-]}

\item{Kp}{subcompartment/water partitioning coefficient [-]}

\item{Matrix}{type of compartment}

\item{rhoMatrix}{density of the matrix [-]}
}
\value{
Kscompw
}
\description{
Dimensionless soil/water partition coefficient specific to subcompartment
}
