% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Ksdcompw.R
\name{Ksdcompw}
\alias{Ksdcompw}
\title{Dimensionless sed/water PARTITION COEFFICIENT for molecular species specific to sediment compartment}
\usage{
Ksdcompw(FRACw, FRACs, Kp, all.rhoMatrix, Matrix)
}
\arguments{
\item{FRACw}{fraction of water in compartment [-]}

\item{FRACs}{fraction of soil in compartment [-]}

\item{Kp}{general sediment water partitioning coefficient [-]}

\item{Matrix}{type of compartment considered}

\item{rhoMatrix}{density of the matrix [kg m-3]}
}
\value{
Ksdcompw
}
\description{
Dimensionless sediment/water partitioning coefficient dependent on the compartment
}
