% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_kSense.R
\name{kSense}
\alias{kSense}
\title{kSense}
\usage{
kSense(ParentModule, knamesdist, samplesize)
}
\arguments{
\item{ParentModule}{SBCore}

\item{knamesdist}{description}

\item{samplesize}{description}
}
\value{
States (i) (=mass)
}
\description{
sensitivity Mass 4 kaas,
#logged sd / intervals ??
 by changing kaas what is the change in Mass
}
