% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Ksw.alt.R
\name{Ksw.alt soil water partitioning coefficient for colloids}
\alias{Ksw.alt soil water partitioning coefficient for colloids}
\alias{Ksw.alt}
\title{Ksw.alt}
\usage{
Ksw.alt(Kow, pKa, CorgStandard, ChemClass, a, b, all.rhoMatrix, KswDorC)
}
\arguments{
\item{Kow}{octanol water partitioning coefficient [-]}

\item{pKa}{Dissociation constant of (conjugated) acid (default = 7}

\item{CorgStandard}{Standard mass FRACTION organic carbon in soil/sediment [-]}

\item{ChemClass}{Class of chemical, see QSAR table (REACH, 2012)}

\item{a}{see QSAR table}

\item{b}{see QSAR table}

\item{KswDorC}{soil water partitioning coefficient for colloids [-]}

\item{rhoMatrix}{density of the matrix [kg/m3]}
}
\description{
Ksw.alt
}
