% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_MasConc_Otherparticle.R
\name{MasConc_Otherparticle}
\alias{MasConc_Otherparticle}
\title{MasConc_Otherparticle}
\usage{
MasConc_Otherparticle(mConcSusp, mConcCol, SpeciesName)
}
\arguments{
\item{mConcSusp}{concentration of suspended matter in considered matrix}

\item{mConcCol}{concentration of colloids in considered matrix}

\item{SpeciesName}{name of the considered species}
}
\value{
MasConc_Otherparticle
}
\description{
MasConc_Otherparticle
}
