% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Mass2Conc.R
\name{Mass2Conc}
\alias{Mass2Conc}
\title{calculate concentration from mass}
\usage{
Mass2Conc(EqMass, Volume, Matrix, all.rhoMatrix, Fracs, Fracw)
}
\arguments{
\item{Volume}{of the Box (state) [m3]}

\item{Fracs}{fraction of solids in the compartment [-]}

\item{Fracw}{fraction of water in compartment [-]}

\item{eqMass}{mass, as returned by solvers [kg]}

\item{rhoMatrix}{density of the matrix [kg m-3]}
}
\value{
Concentration, kg/kg wet for soils, kg/kg wet for sediment, otherwise kg/m3
}
\description{
calculates Concentrations from mass per Box [g/kg] or [kg/m3]
}
