% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Mass2Conc.R
\name{Mass2Conc}
\alias{Mass2Conc}
\title{calculate concentration from mass}
\usage{
Mass2Conc(EqMass, Volume, Matrix, all.rhoMatrix, Fracs, Fracw)
}
\arguments{
\item{Volume}{of the Box (state)}

\item{eqMass}{mass, as returned by solvers}
}
\value{
Concentration, kg/kg wet for soils, kg/kg wet for sediment, otherwise kg/m3
}
\description{
calculates Concentrations from mass per Box
}
