% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_MTC_2a.R
\name{MTC_2a}
\alias{MTC_2a}
\title{PARTIAL MASS TRANSFER COEFFICIENT water and soil to air}
\usage{
MTC_2a(
  WINDspeed,
  MW,
  Tempfactor,
  KdegDorC,
  Matrix,
  SpeciesName,
  ScaleName,
  SubCompartName
)
}
\arguments{
\item{WINDspeed}{Windspeed in compartment/scale [m.s-1]}

\item{MW}{Molecular weight of compound [g.mol-1]}

\item{from.Matrix}{Matrix/compartment from which the relevant process is taking place}
}
\description{
Partial mass transfer coefficient from water and soil to air
}
