% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_MTC_2s.R
\name{MTC_2s}
\alias{MTC_2s}
\title{PARTIAL MASS TRANSFER COEFFICIENT air to soil}
\usage{
MTC_2s(Mackay1, Mackay2, Matrix)
}
\arguments{
\item{Mackay1}{constant described by Mackay (2001) [m/s] # data in m/d (units conversion)}

\item{Mackay2}{constant described by Mackay (2001) [-]}
}
\description{
PARTIAL MASS TRANSFER COEFFICIENT air to soil
}
