% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_MTC_2sd.R
\name{MTC_2sd}
\alias{MTC_2sd}
\title{PARTIAL MASS TRANSFER COEFFICIENT water to sediment}
\usage{
MTC_2sd(kwsd.water, Matrix, SubCompartName)
}
\arguments{
\item{kwsd.water}{Constant water side pMTC to sediment [m/s-1]}
}
\description{
PARTIAL MASS TRANSFER COEFFICIENT water to sediment
}
