% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_rad_species.R
\name{rad_species}
\alias{rad_species}
\title{Radius of species}
\usage{
rad_species(
  SpeciesName,
  SubCompartName,
  RadCOL,
  RadCP,
  RadNuc,
  RadS,
  NumConcNuc,
  NumConcAcc,
  Df
)
}
\arguments{
\item{SpeciesName}{species considered}

\item{SubCompartName}{subcompartment considered}

\item{RadCOL}{Accumulation mode aerosol particle radius [m]}

\item{RadCP}{coarse particulate mode aerosol particle radius [m]}

\item{RadNuc}{Nucleation mode aerosol particle radius [m]}

\item{RadS}{nanoparticle radius [m]}

\item{NumConcNuc}{Number concentration of Nucleation mode aerosol particles [#/m3]}

\item{NumConcAcc}{Number concentration of Accumulation mode aerosol particles [#/m3]}

\item{Df}{fractal dimension of combined heteroagglomerate [-]}

\item{NaturalRad}{natural particle radius all but small in air [m]}
}
\value{
rad_species, Approach to calculate the radius of small heteroagglomerates in air/water [m]
}
\description{
Calculate radius of heteroagglomerates [m]
}
