% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_RainOnFreshwater.R
\name{RainOnFreshwater}
\alias{RainOnFreshwater}
\title{RainOnFreshwater}
\usage{
RainOnFreshwater(RAINrate, Area, SubCompartName)
}
\arguments{
\item{RAINrate}{m.s-1}

\item{Area}{in m2}

\item{SubCompartName}{#only for lake/rivers}
}
\value{
waterflow of rain directly on lake/river
and continental being a part of Moderate (/ Tropic)
}
\description{
RainOnFreshwater
}
