% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_SBsolve.R
\name{SBsolve}
\alias{SBsolve}
\title{SBsolve}
\usage{
SBsolve(ParentModule, tmax = 1e+10, nTIMES = 100)
}
\arguments{
\item{tmax}{}
}
\value{
see: ode()
}
\description{
solve system of States for 1rst order k(i,j) and emissions
}
