% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6SBstates.R
\name{SBstates}
\alias{SBstates}
\title{states}
\description{
object in both SBcore as in SolverModule (slimmed down to existing kaas).
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{asDataFrame}}{convienent returning states in a data.frame}

\item{\code{nStates}}{just 4 convenience}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SBstates-new}{\code{SBstates$new()}}
\item \href{#method-SBstates-findState}{\code{SBstates$findState()}}
\item \href{#method-SBstates-sortFactors}{\code{SBstates$sortFactors()}}
\item \href{#method-SBstates-findDim}{\code{SBstates$findDim()}}
\item \href{#method-SBstates-clone}{\code{SBstates$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-new"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-new}{}}}
\subsection{Method \code{new()}}{
init, based on
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$new(StatesAsDataFrame)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{StatesAsDataFrame}}{dataframe of states with columns The3D Abbr and counter i}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-findState"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-findState}{}}}
\subsection{Method \code{findState()}}{
map vector abbr for scale species subcompart to index in State
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$findState(abbr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{abbr}}{Oldschool abbreviation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of indices
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-sortFactors"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-sortFactors}{}}}
\subsection{Method \code{sortFactors()}}{
change any of the 3 Dim into factors and sort a data.frame accordingly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$sortFactors(aDFwithD)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aDFwithD}}{a data.frame to sort}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
sorted dataframe with dimensions from 3D as factor with levels/labels according data sheets
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-findDim"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-findDim}{}}}
\subsection{Method \code{findDim()}}{
find an element in any dimension, or the name of a dimension NOT FINISHED
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$findDim(aDimension)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aDimension}}{a dimension or an element therein}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list(the found dimension = the member) or list(the found dimension = dimension). 
The found dimension is 1 of the 3dim
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SBstates-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SBstates-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBstates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
