% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6SolverModule.R
\name{SolverModule}
\alias{SolverModule}
\title{SolverModule}
\value{
dm (i) = change in mass as list
}
\description{
wrapper for your solver function and collection of general methods needed for most solvers, 
like preparing kaas matrix and matching emissions vector and
the general ODE for simplebox
}
\section{Super class}{
\code{\link[sboo:CalcGraphModule]{sboo::CalcGraphModule}} -> \code{SolverModule}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{needVars}}{getter for parameters, derived from the defining function}

\item{\code{states}}{injected from States}

\item{\code{emissions}}{guess what?}

\item{\code{SB.k}}{r.o. matrix of k's, after preparing, mostly abuse of identity for removal processes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SolverModule-execute}{\code{SolverModule$execute()}}
\item \href{#method-SolverModule-PrepKaasM}{\code{SolverModule$PrepKaasM()}}
\item \href{#method-SolverModule-PrepemisV}{\code{SolverModule$PrepemisV()}}
\item \href{#method-SolverModule-SimpleBoxODE}{\code{SolverModule$SimpleBoxODE()}}
\item \href{#method-SolverModule-DiffSB.K}{\code{SolverModule$DiffSB.K()}}
\item \href{#method-SolverModule-clone}{\code{SolverModule$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="sboo" data-topic="CalcGraphModule" data-id="initialize"><a href='../../sboo/html/CalcGraphModule.html#method-CalcGraphModule-initialize'><code>sboo::CalcGraphModule$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolverModule-execute"></a>}}
\if{latex}{\out{\hypertarget{method-SolverModule-execute}{}}}
\subsection{Method \code{execute()}}{
prepare kaas for matrix calculations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolverModule$execute(needdebug = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{needdebug}}{if T, the solver defining function execute in debugmode}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolverModule-PrepKaasM"></a>}}
\if{latex}{\out{\hypertarget{method-SolverModule-PrepKaasM}{}}}
\subsection{Method \code{PrepKaasM()}}{
prepare kaas for matrix calculations
1 Convert relational i,j,kaas into a matrix (matrify, pivot..)
including addition of kaas with identical (i,j)
2 The kaas are only describing removal, the where-to needs to be
added to the from-masses by putting it to into the diagonal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolverModule$PrepKaasM(kaas = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{kaas}}{k's}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
matrix with kaas; ready to go
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolverModule-PrepemisV"></a>}}
\if{latex}{\out{\hypertarget{method-SolverModule-PrepemisV}{}}}
\subsection{Method \code{PrepemisV()}}{
sync emissions as relational table with states into vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolverModule$PrepemisV(emissions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{emissions}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
matrix with kaas; ready to go
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolverModule-SimpleBoxODE"></a>}}
\if{latex}{\out{\hypertarget{method-SolverModule-SimpleBoxODE}{}}}
\subsection{Method \code{SimpleBoxODE()}}{
basic ODE function for simplebox; the function for the ode-call; 
see desolve/rootsolve packages
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolverModule$SimpleBoxODE(t, m, parms)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{time (vector ?)}

\item{\code{m}}{(i) = initial mass}

\item{\code{parms}}{= kaas, emissions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolverModule-DiffSB.K"></a>}}
\if{latex}{\out{\hypertarget{method-SolverModule-DiffSB.K}{}}}
\subsection{Method \code{DiffSB.K()}}{
diff between kaas in this and k's in OtherSB.K
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolverModule$DiffSB.K(OtherSB.K, tiny = 1e-20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{OtherSB.K}}{the 'other' kaas}

\item{\code{tiny}}{(epsilon) permitted rounding error (we might be dealing with excel ! :( )}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SolverModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SolverModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolverModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
