% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_KdegDorC.R
\name{v_KdegDorC}
\alias{v_KdegDorC}
\alias{KdegDorC}
\title{Degradation rate constant measured or calculated}
\usage{
KdegDorC(
  kdeg,
  C.OHrad.n,
  k0.OHrad,
  Ea.OHrad,
  T25,
  Q.10,
  KswDorC,
  Biodeg,
  CorgStandard,
  rhoMatrix,
  Matrix,
  SpeciesName
)
}
\arguments{
\item{C.OHrad.n}{?}

\item{k0.OHrad}{?}

\item{Ea.OHrad}{?}

\item{T25}{= 298 K}

\item{FRingas}{see}

\item{kdeg.air}{kdeg.water kdeg.water}

\item{C.OHrad}{?}

\item{Temp}{[K]}
}
\value{
Degradation rate constant for molecular species
}
\description{
calculate k for degradation;
}
