% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_KdegDorC.R
\name{v_KdegDorC}
\alias{v_KdegDorC}
\alias{KdegDorC}
\title{Degradation rate constant measured or calculated}
\usage{
KdegDorC(
  kdeg,
  C.OHrad.n,
  k0.OHrad,
  Ea.OHrad,
  T25,
  Q.10,
  KswDorC,
  Biodeg,
  CorgStandard,
  rhoMatrix,
  Matrix,
  SpeciesName
)
}
\arguments{
\item{kdeg}{degradation rate (as input) [s-1]}

\item{C.OHrad.n}{general OH radical concentration, based on Wania & Daly (2002) [mol m-3]}

\item{k0.OHrad}{frequency factor of the OH radical reaction [m3 s-1]}

\item{Ea.OHrad}{activation energy OH radical reaction [J mol-1]}

\item{T25}{298K [K]}

\item{Q.10}{rate increase factor per 10C [-]}

\item{KswDorC}{calculated soil water partitioning coefficient  [-]}

\item{CorgStandard}{Standard mass fraction organic carbon in soil/sediment [-]}

\item{rhoMatrix}{density of the matrix}

\item{Matrix}{compartment type considered}

\item{SpeciesName}{species name considered}

\item{C.OHrad}{OH radical concentration specific to compartment, based on Wania & Daly (2002) [mol m-3]}

\item{BioDeg}{biodegradability test result [-]}
}
\value{
Degradation rate constant for molecular species
}
\description{
calculate k for degradation for particulates and molecules.
if no specific value for kdeg is available it is scaled into categories 
as defined in Technical Guidance Document on risk assessment in support of
Commission Directive 93/67/EEC (European Commission, 2003)
}
