% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_OtherkAir.R
\name{v_OtherkAir}
\alias{v_OtherkAir}
\alias{OtherkAir}
\title{v_OtherkAir}
\usage{
OtherkAir(all.kaas, ScaleName, SpeciesName)
}
\arguments{
\item{all.kaas}{sneaky way to obtain other loss processes from air}
}
\description{
v_OtherkAir
}
