% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_Volume.R
\name{Volume}
\alias{Volume}
\title{Volume of the SubCompartment}
\usage{
Volume(VertDistance, Area, FRACcldw, SubCompartName)
}
\arguments{
\item{VertDistance}{height or depth for subcompartments below the horizon [m]}

\item{Area}{area of the compartment [m2]}

\item{FRACcldw}{fraction of cloudwater [-]}

\item{SubCompartName}{subcompartment considered}
}
\value{
Volume
}
\description{
Volume of the SubCompartment
}
