% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_Advection_Air.R
\name{x_Advection_air}
\alias{x_Advection_air}
\alias{x_Advection_Air}
\title{Advection of fluxes of air between scales}
\usage{
x_Advection_Air(all.AirFlow, from.ScaleName, to.ScaleName)
}
\arguments{
\item{all.AirFlow}{to pick for the "from" scale}

\item{from.ScaleName}{name of the scale of the source of the process}

\item{to.ScaleName}{name of the scale of the destination of the process}
}
\value{
x.Advection_Air, the flow
}
\description{
Calculation using AirFlow, but adjust to mass balance volume flows
}
