% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ContRiver2Reg.R
\name{x_ContRiver2Reg}
\alias{x_ContRiver2Reg}
\title{Continental River 2 Regional}
\usage{
x_ContRiver2Reg(
  ScaleName,
  SubCompartName,
  all.Runoff,
  RainOnFreshwater,
  dischargeFRAC,
  LakeFracRiver
)
}
\arguments{
\item{dischargeFRAC}{fraction of ...}

\item{x_Runoff}{runoff flux}
}
\value{
River Discharge for scale Continental
}
\description{
Continental River 2 Regional
}
