% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ContRiver2Reg.R
\name{x_ContRiver2Reg}
\alias{x_ContRiver2Reg}
\title{Continental surface water to Regional}
\usage{
x_ContRiver2Reg(
  ScaleName,
  SubCompartName,
  all.Runoff,
  RainOnFreshwater,
  dischargeFRAC,
  LakeFracRiver
)
}
\arguments{
\item{ScaleName}{Name of the relevant scale}

\item{SubCompartName}{Name of the relevant sub-compartment}

\item{RainOnFreshwater}{}

\item{dischargeFRAC}{fraction of discharge that is going from continental to regional surface water}

\item{LakeFracRiver}{description}

\item{Runoff}{runoff flow from soil to river}
}
\value{
River Discharge for scale Continental
}
\description{
Continental surface water to Regional
}
