% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ContSea2Reg.R
\name{x_ContSea2Reg}
\alias{x_ContSea2Reg}
\title{ContSea2Reg}
\usage{
x_ContSea2Reg(all.x_RiverDischarge, LakeFracRiver)
}
\arguments{
\item{LakeFracRiver}{see ppt}

\item{RiverDischarge}{see ppt}
}
\value{
additional flux from continental sea to regional
}
\description{
there is an additional flux from continental sea to regional to compensate from the LakeFracRiver
}
