% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_FromModerate2TropWater.R
\name{x_FromModerate2TropWater}
\alias{x_FromModerate2TropWater}
\title{Mixing of upper and deeper sea water layers (w2 and w3)}
\usage{
x_FromModerate2TropWater(OceanCurrent, SubCompartName, ScaleName)
}
\arguments{
\item{OceanCurrent}{[m3.s-1]}

\item{SubCompartName}{name of the subcompartment of the box at hand}

\item{ScaleName}{name of the scale of the box at hand}
}
\value{
Advection sea - deepocean
}
\description{
Mixing of upper and deeper sea water layers (w2 and w3)
}
