% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_RiverDischarge.R
\name{x_RiverDischarge}
\alias{x_RiverDischarge}
\title{RiverDischarge}
\usage{
x_RiverDischarge(
  all.Runoff,
  RainOnFreshwater,
  dischargeFRAC,
  all.x_ContRiver2Reg,
  ScaleName,
  SubCompartName
)
}
\arguments{
\item{dischargeFRAC}{Fraction discharge between freshwater compartment (w1) at regional and continental scale}

\item{RainRunoff}{see ppt}
}
\value{
River Discharge
}
\description{
RiverDischarge
}
