% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ToModerateWater.R
\name{x_ToModerateWater}
\alias{x_ToModerateWater}
\title{Mixing of upper and deeper sea water layers (w2 and w3)}
\usage{
x_ToModerateWater(
  Volume,
  TAUsea,
  all.x_RegSea2Cont,
  OceanCurrent,
  SubCompartName,
  ScaleName
)
}
\arguments{
\item{OceanCurrent}{[m3.s-1]}

\item{SubCompartName}{name of the subcompartment of the box at hand}

\item{ScaleName}{name of the scale of the box at hand}
}
\value{
Advection sea - deepocean
}
\description{
Mixing of upper and deeper sea water layers (w2 and w3)
}
