% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ToModerateWater.R
\name{x_ToModerateWater}
\alias{x_ToModerateWater}
\title{Ocean currents flowing to Moderate scale (w2 and w3)}
\usage{
x_ToModerateWater(
  Volume,
  TAUsea,
  all.x_RegSea2Cont,
  OceanCurrent,
  SubCompartName,
  ScaleName
)
}
\arguments{
\item{Volume}{Volume of compartment [m3]}

\item{TAUsea}{Residence time of water in sea - scale variable [s]}

\item{all.x_RegSea2Cont}{All flows from regional to continental seawater [s-1]}

\item{OceanCurrent}{Global ocean circulation current [m3.s-1]}

\item{SubCompartName}{Name of the subcompartment of the box at hand}

\item{ScaleName}{Name of the scale of the box at hand}
}
\value{
Water flow to Moderate scale surface and deepocean waters [m3 s-1]
}
\description{
Advection of continental sea, Tropic and Arctic sea compartments to sea and deepocean at Moderate scale.
}
